+++ src/main/scala/org/globalforestwatch/treecoverloss/Adm2DF.scala
@@ -1,0 +1,189 @@
+package org.globalforestwatch.treecoverloss
+
+import org.apache.spark.sql._
+import org.apache.spark.sql.functions._
+import com.github.mrpowers.spark.daria.sql.DataFrameHelpers._
+import org.globalforestwatch.treecoverloss.TreeLossDFHelpers.windowSum
+
+object Adm2DF {
+
+  val spark: SparkSession = TreeLossSparkSession().spark
+  import spark.implicits._
+
+  def joinExtent2010(extent2010DF: DataFrame)(df: DataFrame): DataFrame = {
+    df.join(
+      extent2010DF,
+      $"m_feature_id" <=> $"feature_id" &&
+        $"m_layers" <=> $"layers" &&
+        $"m_threshold" <=> $"threshold",
+      "left_outer"
+    )
+  }
+
+  def unpackFeautureIDLayers(df: DataFrame): DataFrame = {
+    df.select(
+      $"feature_id.iso" as "iso",
+      $"feature_id.adm1" as "adm1",
+      $"feature_id.adm2" as "adm2",
+      $"threshold",
+      $"layers.drivers" as "tcs",
+      $"layers.globalLandCover" as "global_land_cover",
+      $"layers.primaryForest" as "primary_forest",
+      $"layers.idnPrimaryForest" as "idn_primary_forest",
+      $"layers.erosion" as "erosion",
+      $"layers.biodiversitySignificance" as "biodiversity_significance",
+      $"layers.biodiversityIntactness" as "biodiversity_intactness",
+      $"layers.wdpa" as "wdpa",
+      $"layers.aze" as "aze",
+      $"layers.plantations" as "plantations",
+      $"layers.riverBasins" as "river_basin",
+      $"layers.ecozones" as "ecozone",
+      $"layers.urbanWatersheds" as "urban_watershed",
+      $"layers.mangroves1996" as "mangroves_1996",
+      $"layers.mangroves2016" as "mangroves_2016",
+      $"layers.waterStress" as "water_stress",
+      $"layers.intactForestLandscapes" as "ifl",
+      $"layers.endemicBirdAreas" as "endemic_bird_area",
+      $"layers.tigerLandscapes" as "tiger_cl",
+      $"layers.landmark" as "landmark",
+      $"layers.landRights" as "land_right",
+      $"layers.keyBiodiversityAreas" as "kba",
+      $"layers.mining" as "mining",
+      $"layers.rspo" as "rspo",
+      $"layers.peatlands" as "idn_mys_peatlands",
+      $"layers.oilPalm" as "oil_palm",
+      $"layers.idnForestMoratorium" as "idn_forest_moratorium",
+      $"layers.idnLandCover" as "idn_land_cover",
+      $"layers.mexProtectedAreas" as "mex_protected_areas",
+      $"layers.mexPaymentForEcosystemServices" as "mex_pes",
+      $"layers.mexForestZoning" as "mex_forest_zoning",
+      $"layers.perProductionForest" as "per_production_forest",
+      $"layers.perProtectedAreas" as "per_protected_area",
+      $"layers.perForestConcessions" as "per_forest_concession",
+      $"layers.braBiomes" as "bra_biomes",
+      $"layers.woodFiber" as "wood_fiber",
+      $"layers.resourceRights" as "resource_right",
+      $"layers.logging" as "managed_forests",
+      $"layers.oilGas" as "oil_gas",
+      $"totalarea",
+      $"extent2000",
+      $"extent2010",
+      $"total_gain",
+      $"total_biomass",
+      $"avg_biomass_per_ha",
+      $"total_co2",
+      $"total_mangrove_biomass",
+      $"avg_mangrove_biomass_per_ha",
+      $"total_mangrove_co2",
+      $"year_2001",
+      $"area_loss_2001",
+      $"biomass_loss_2001",
+      $"carbon_emissions_2001",
+      $"mangrove_biomass_loss_2001",
+      $"mangrove_carbon_emissions_2001",
+      $"year_2002",
+      $"area_loss_2002",
+      $"biomass_loss_2002",
+      $"carbon_emissions_2002",
+      $"mangrove_biomass_loss_2002",
+      $"mangrove_carbon_emissions_2002",
+      $"year_2003",
+      $"area_loss_2003",
+      $"biomass_loss_2003",
+      $"carbon_emissions_2003",
+      $"mangrove_biomass_loss_2003",
+      $"mangrove_carbon_emissions_2003",
+      $"year_2004",
+      $"area_loss_2004",
+      $"biomass_loss_2004",
+      $"carbon_emissions_2004",
+      $"mangrove_biomass_loss_2004",
+      $"mangrove_carbon_emissions_2004",
+      $"year_2005",
+      $"area_loss_2005",
+      $"biomass_loss_2005",
+      $"carbon_emissions_2005",
+      $"mangrove_biomass_loss_2005",
+      $"mangrove_carbon_emissions_2005",
+      $"year_2006",
+      $"area_loss_2006",
+      $"biomass_loss_2006",
+      $"carbon_emissions_2006",
+      $"mangrove_biomass_loss_2006",
+      $"mangrove_carbon_emissions_2006",
+      $"year_2007",
+      $"area_loss_2007",
+      $"biomass_loss_2007",
+      $"carbon_emissions_2007",
+      $"mangrove_biomass_loss_2007",
+      $"mangrove_carbon_emissions_2007",
+      $"year_2008",
+      $"area_loss_2008",
+      $"biomass_loss_2008",
+      $"carbon_emissions_2008",
+      $"mangrove_biomass_loss_2008",
+      $"mangrove_carbon_emissions_2008",
+      $"year_2009",
+      $"area_loss_2009",
+      $"biomass_loss_2009",
+      $"carbon_emissions_2009",
+      $"mangrove_biomass_loss_2009",
+      $"mangrove_carbon_emissions_2009",
+      $"year_2010",
+      $"area_loss_2010",
+      $"biomass_loss_2010",
+      $"carbon_emissions_2010",
+      $"mangrove_biomass_loss_2010",
+      $"mangrove_carbon_emissions_2010",
+      $"year_2011",
+      $"area_loss_2011",
+      $"biomass_loss_2011",
+      $"carbon_emissions_2011",
+      $"mangrove_biomass_loss_2011",
+      $"mangrove_carbon_emissions_2011",
+      $"year_2012",
+      $"area_loss_2012",
+      $"biomass_loss_2012",
+      $"carbon_emissions_2012",
+      $"mangrove_biomass_loss_2012",
+      $"mangrove_carbon_emissions_2012",
+      $"year_2013",
+      $"area_loss_2013",
+      $"biomass_loss_2013",
+      $"carbon_emissions_2013",
+      $"mangrove_biomass_loss_2013",
+      $"mangrove_carbon_emissions_2013",
+      $"year_2014",
+      $"area_loss_2014",
+      $"biomass_loss_2014",
+      $"carbon_emissions_2014",
+      $"mangrove_biomass_loss_2014",
+      $"mangrove_carbon_emissions_2014",
+      $"year_2015",
+      $"area_loss_2015",
+      $"biomass_loss_2015",
+      $"carbon_emissions_2015",
+      $"mangrove_biomass_loss_2015",
+      $"mangrove_carbon_emissions_2015",
+      $"year_2016",
+      $"area_loss_2016",
+      $"biomass_loss_2016",
+      $"carbon_emissions_2016",
+      $"mangrove_biomass_loss_2016",
+      $"mangrove_carbon_emissions_2016",
+      $"year_2017",
+      $"area_loss_2017",
+      $"biomass_loss_2017",
+      $"carbon_emissions_2017",
+      $"mangrove_biomass_loss_2017",
+      $"mangrove_carbon_emissions_2017",
+      $"year_2018",
+      $"area_loss_2018",
+      $"biomass_loss_2018",
+      $"carbon_emissions_2018",
+      $"mangrove_biomass_loss_2018",
+      $"mangrove_carbon_emissions_2018"
+    )
+  }
+
+}
