+++ src/main/scala/org/globalforestwatch/treecoverloss/TreeLossSummaryMain.scala
@@ -5,14 +5,11 @@ import org.apache.log4j.Logger
 import org.apache.spark._
 import org.apache.spark.rdd._
 import org.apache.spark.sql._
-import org.apache.spark.sql.functions._
 import cats.implicits._
 import geotrellis.vector.io.wkb.WKB
 import geotrellis.vector.{Feature, Geometry}
 import java.time.format.DateTimeFormatter
 import java.time.LocalDateTime
-import TreeLossDFHelpers._
-
 
 object TreeLossSummaryMain
   extends CommandApp(
@@ -74,7 +71,7 @@ object TreeLossSummaryMain
          iso,
          admin1,
          admin2) =>
-          val spark: SparkSession = TreeLossSparkSession().spark
+          val spark: SparkSession = TreeLossSparkSession.spark
           import spark.implicits._
 
           // ref: https://github.com/databricks/spark-csv
@@ -205,7 +202,9 @@ object TreeLossSummaryMain
                 "year_data"
               )
 
-          val runOutputUrl = outputUrl + "/treecoverloss_" + DateTimeFormatter.ofPattern("yyyyMMdd_HHmm").format(LocalDateTime.now)
+          val runOutputUrl = outputUrl + "/treecoverloss_" + DateTimeFormatter
+            .ofPattern("yyyyMMdd_HHmm")
+            .format(LocalDateTime.now)
 
           //          val outputPartitionCount =
           //            maybeOutputPartitions.getOrElse(featureRDD.getNumPartitions)
@@ -213,886 +212,24 @@ object TreeLossSummaryMain
           summaryDF.repartition($"feature_id", $"threshold_2000")
           summaryDF.cache()
 
-
-
-
-          //          val WindowPartitionOrder = Window
-          //            .partitionBy($"m_feature_id", $"m_layers")
-          //            .orderBy($"m_threshold".desc)
-          //
-          //          def windowSum(e: Column): Column = sum(e).over(WindowPartitionOrder)
-          //          def windowSum(columnName: String): Column = windowSum(col(columnName))
-
-          val thresholdDF =
-            Seq(0, 10, 15, 20, 25, 30, 50, 75).toDF("threshold")
-
-          val masterDF = summaryDF
-            .groupBy("feature_id", "layers")
-            .agg(sum("area") as "totalarea")
-            .crossJoin(thresholdDF)
-            .select(
-              $"feature_id" as "m_feature_id",
-              $"layers" as "m_layers",
-              $"threshold" as "m_threshold",
-              $"totalarea"
-            )
+          val masterDF = summaryDF.transform(MasterDF.expandByThreshold)
 
           masterDF.cache()
 
-          val lookup2010 = Map("threshold_2010" -> "threshold")
-
           val extent2010DF = summaryDF
-            .select(
-              summaryDF.columns
-                .map(c => col(c).as(lookup2010.getOrElse(c, c))): _*
-            )
-            .groupBy("feature_id", "layers", "threshold")
-            .agg(sum("area") as "area")
-            .join(
-              masterDF,
-              $"m_feature_id" <=> $"feature_id" &&
-                $"m_layers" <=> $"layers" &&
-                $"m_threshold" <=> $"threshold",
-              "right_outer"
-            )
-            .na
-            .fill(0.0, Seq("area"))
-            .select(
-              $"m_feature_id",
-              $"m_layers",
-              $"m_threshold",
-              windowSum("area") as "extent2010"
-            )
-
+            .transform(Extent2010DF.sumArea)
+            .transform(Extent2010DF.joinMaster(masterDF))
+            .transform(Extent2010DF.aggregateByThreshold)
 
           val annualLossDF = summaryDF
-            .select(
-              $"feature_id",
-              $"threshold_2000" as "threshold",
-              $"layers",
-              $"area",
-              $"gain",
-              $"biomass",
-              $"biomass_per_ha",
-              $"co2",
-              $"mangrove_biomass",
-              $"mangrove_biomass_per_ha",
-              $"mangrove_co2",
-              'year_data.getItem(0).getItem("year") as "year_2001",
-              'year_data.getItem(1).getItem("year") as "year_2002",
-              'year_data.getItem(2).getItem("year") as "year_2003",
-              'year_data.getItem(3).getItem("year") as "year_2004",
-              'year_data.getItem(4).getItem("year") as "year_2005",
-              'year_data.getItem(5).getItem("year") as "year_2006",
-              'year_data.getItem(6).getItem("year") as "year_2007",
-              'year_data.getItem(7).getItem("year") as "year_2008",
-              'year_data.getItem(8).getItem("year") as "year_2009",
-              'year_data.getItem(9).getItem("year") as "year_2010",
-              'year_data.getItem(10).getItem("year") as "year_2011",
-              'year_data.getItem(11).getItem("year") as "year_2012",
-              'year_data.getItem(12).getItem("year") as "year_2013",
-              'year_data.getItem(13).getItem("year") as "year_2014",
-              'year_data.getItem(14).getItem("year") as "year_2015",
-              'year_data.getItem(15).getItem("year") as "year_2016",
-              'year_data.getItem(16).getItem("year") as "year_2017",
-              'year_data.getItem(17).getItem("year") as "year_2018",
-              'year_data.getItem(0).getItem("area_loss") as "area_loss_2001",
-              'year_data.getItem(1).getItem("area_loss") as "area_loss_2002",
-              'year_data.getItem(2).getItem("area_loss") as "area_loss_2003",
-              'year_data.getItem(3).getItem("area_loss") as "area_loss_2004",
-              'year_data.getItem(4).getItem("area_loss") as "area_loss_2005",
-              'year_data.getItem(5).getItem("area_loss") as "area_loss_2006",
-              'year_data.getItem(6).getItem("area_loss") as "area_loss_2007",
-              'year_data.getItem(7).getItem("area_loss") as "area_loss_2008",
-              'year_data.getItem(8).getItem("area_loss") as "area_loss_2009",
-              'year_data.getItem(9).getItem("area_loss") as "area_loss_2010",
-              'year_data.getItem(10).getItem("area_loss") as "area_loss_2011",
-              'year_data.getItem(11).getItem("area_loss") as "area_loss_2012",
-              'year_data.getItem(12).getItem("area_loss") as "area_loss_2013",
-              'year_data.getItem(13).getItem("area_loss") as "area_loss_2014",
-              'year_data.getItem(14).getItem("area_loss") as "area_loss_2015",
-              'year_data.getItem(15).getItem("area_loss") as "area_loss_2016",
-              'year_data.getItem(16).getItem("area_loss") as "area_loss_2017",
-              'year_data.getItem(17).getItem("area_loss") as "area_loss_2018",
-              'year_data
-                .getItem(0)
-                .getItem("biomass_loss") as "biomass_loss_2001",
-              'year_data
-                .getItem(1)
-                .getItem("biomass_loss") as "biomass_loss_2002",
-              'year_data
-                .getItem(2)
-                .getItem("biomass_loss") as "biomass_loss_2003",
-              'year_data
-                .getItem(3)
-                .getItem("biomass_loss") as "biomass_loss_2004",
-              'year_data
-                .getItem(4)
-                .getItem("biomass_loss") as "biomass_loss_2005",
-              'year_data
-                .getItem(5)
-                .getItem("biomass_loss") as "biomass_loss_2006",
-              'year_data
-                .getItem(6)
-                .getItem("biomass_loss") as "biomass_loss_2007",
-              'year_data
-                .getItem(7)
-                .getItem("biomass_loss") as "biomass_loss_2008",
-              'year_data
-                .getItem(8)
-                .getItem("biomass_loss") as "biomass_loss_2009",
-              'year_data
-                .getItem(9)
-                .getItem("biomass_loss") as "biomass_loss_2010",
-              'year_data
-                .getItem(10)
-                .getItem("biomass_loss") as "biomass_loss_2011",
-              'year_data
-                .getItem(11)
-                .getItem("biomass_loss") as "biomass_loss_2012",
-              'year_data
-                .getItem(12)
-                .getItem("biomass_loss") as "biomass_loss_2013",
-              'year_data
-                .getItem(13)
-                .getItem("biomass_loss") as "biomass_loss_2014",
-              'year_data
-                .getItem(14)
-                .getItem("biomass_loss") as "biomass_loss_2015",
-              'year_data
-                .getItem(15)
-                .getItem("biomass_loss") as "biomass_loss_2016",
-              'year_data
-                .getItem(16)
-                .getItem("biomass_loss") as "biomass_loss_2017",
-              'year_data
-                .getItem(17)
-                .getItem("biomass_loss") as "biomass_loss_2018",
-              'year_data
-                .getItem(0)
-                .getItem("carbon_emissions") as "carbon_emissions_2001",
-              'year_data
-                .getItem(1)
-                .getItem("carbon_emissions") as "carbon_emissions_2002",
-              'year_data
-                .getItem(2)
-                .getItem("carbon_emissions") as "carbon_emissions_2003",
-              'year_data
-                .getItem(3)
-                .getItem("carbon_emissions") as "carbon_emissions_2004",
-              'year_data
-                .getItem(4)
-                .getItem("carbon_emissions") as "carbon_emissions_2005",
-              'year_data
-                .getItem(5)
-                .getItem("carbon_emissions") as "carbon_emissions_2006",
-              'year_data
-                .getItem(6)
-                .getItem("carbon_emissions") as "carbon_emissions_2007",
-              'year_data
-                .getItem(7)
-                .getItem("carbon_emissions") as "carbon_emissions_2008",
-              'year_data
-                .getItem(8)
-                .getItem("carbon_emissions") as "carbon_emissions_2009",
-              'year_data
-                .getItem(9)
-                .getItem("carbon_emissions") as "carbon_emissions_2010",
-              'year_data
-                .getItem(10)
-                .getItem("carbon_emissions") as "carbon_emissions_2011",
-              'year_data
-                .getItem(11)
-                .getItem("carbon_emissions") as "carbon_emissions_2012",
-              'year_data
-                .getItem(12)
-                .getItem("carbon_emissions") as "carbon_emissions_2013",
-              'year_data
-                .getItem(13)
-                .getItem("carbon_emissions") as "carbon_emissions_2014",
-              'year_data
-                .getItem(14)
-                .getItem("carbon_emissions") as "carbon_emissions_2015",
-              'year_data
-                .getItem(15)
-                .getItem("carbon_emissions") as "carbon_emissions_2016",
-              'year_data
-                .getItem(16)
-                .getItem("carbon_emissions") as "carbon_emissions_2017",
-              'year_data
-                .getItem(17)
-                .getItem("carbon_emissions") as "carbon_emissions_2018",
-              'year_data
-                .getItem(0)
-                .getItem("mangrove_biomass_loss") as "mangrove_biomass_loss_2001",
-              'year_data
-                .getItem(1)
-                .getItem("mangrove_biomass_loss") as "mangrove_biomass_loss_2002",
-              'year_data
-                .getItem(2)
-                .getItem("mangrove_biomass_loss") as "mangrove_biomass_loss_2003",
-              'year_data
-                .getItem(3)
-                .getItem("mangrove_biomass_loss") as "mangrove_biomass_loss_2004",
-              'year_data
-                .getItem(4)
-                .getItem("mangrove_biomass_loss") as "mangrove_biomass_loss_2005",
-              'year_data
-                .getItem(5)
-                .getItem("mangrove_biomass_loss") as "mangrove_biomass_loss_2006",
-              'year_data
-                .getItem(6)
-                .getItem("mangrove_biomass_loss") as "mangrove_biomass_loss_2007",
-              'year_data
-                .getItem(7)
-                .getItem("mangrove_biomass_loss") as "mangrove_biomass_loss_2008",
-              'year_data
-                .getItem(8)
-                .getItem("mangrove_biomass_loss") as "mangrove_biomass_loss_2009",
-              'year_data
-                .getItem(9)
-                .getItem("mangrove_biomass_loss") as "mangrove_biomass_loss_2010",
-              'year_data
-                .getItem(10)
-                .getItem("mangrove_biomass_loss") as "mangrove_biomass_loss_2011",
-              'year_data
-                .getItem(11)
-                .getItem("mangrove_biomass_loss") as "mangrove_biomass_loss_2012",
-              'year_data
-                .getItem(12)
-                .getItem("mangrove_biomass_loss") as "mangrove_biomass_loss_2013",
-              'year_data
-                .getItem(13)
-                .getItem("mangrove_biomass_loss") as "mangrove_biomass_loss_2014",
-              'year_data
-                .getItem(14)
-                .getItem("mangrove_biomass_loss") as "mangrove_biomass_loss_2015",
-              'year_data
-                .getItem(15)
-                .getItem("mangrove_biomass_loss") as "mangrove_biomass_loss_2016",
-              'year_data
-                .getItem(16)
-                .getItem("mangrove_biomass_loss") as "mangrove_biomass_loss_2017",
-              'year_data
-                .getItem(17)
-                .getItem("mangrove_biomass_loss") as "mangrove_biomass_loss_2018",
-              'year_data
-                .getItem(0)
-                .getItem("mangrove_carbon_emissions") as "mangrove_carbon_emissions_2001",
-              'year_data
-                .getItem(1)
-                .getItem("mangrove_carbon_emissions") as "mangrove_carbon_emissions_2002",
-              'year_data
-                .getItem(2)
-                .getItem("mangrove_carbon_emissions") as "mangrove_carbon_emissions_2003",
-              'year_data
-                .getItem(3)
-                .getItem("mangrove_carbon_emissions") as "mangrove_carbon_emissions_2004",
-              'year_data
-                .getItem(4)
-                .getItem("mangrove_carbon_emissions") as "mangrove_carbon_emissions_2005",
-              'year_data
-                .getItem(5)
-                .getItem("mangrove_carbon_emissions") as "mangrove_carbon_emissions_2006",
-              'year_data
-                .getItem(6)
-                .getItem("mangrove_carbon_emissions") as "mangrove_carbon_emissions_2007",
-              'year_data
-                .getItem(7)
-                .getItem("mangrove_carbon_emissions") as "mangrove_carbon_emissions_2008",
-              'year_data
-                .getItem(8)
-                .getItem("mangrove_carbon_emissions") as "mangrove_carbon_emissions_2009",
-              'year_data
-                .getItem(9)
-                .getItem("mangrove_carbon_emissions") as "mangrove_carbon_emissions_2010",
-              'year_data
-                .getItem(10)
-                .getItem("mangrove_carbon_emissions") as "mangrove_carbon_emissions_2011",
-              'year_data
-                .getItem(11)
-                .getItem("mangrove_carbon_emissions") as "mangrove_carbon_emissions_2012",
-              'year_data
-                .getItem(12)
-                .getItem("mangrove_carbon_emissions") as "mangrove_carbon_emissions_2013",
-              'year_data
-                .getItem(13)
-                .getItem("mangrove_carbon_emissions") as "mangrove_carbon_emissions_2014",
-              'year_data
-                .getItem(14)
-                .getItem("mangrove_carbon_emissions") as "mangrove_carbon_emissions_2015",
-              'year_data
-                .getItem(15)
-                .getItem("mangrove_carbon_emissions") as "mangrove_carbon_emissions_2016",
-              'year_data
-                .getItem(16)
-                .getItem("mangrove_carbon_emissions") as "mangrove_carbon_emissions_2017",
-              'year_data.getItem(17).getItem("mangrove_carbon_emissions") as "mangrove_carbon_emissions_2018"
-            )
-            .groupBy(
-              $"feature_id",
-              $"threshold",
-              $"layers",
-              $"year_2001",
-              $"year_2002",
-              $"year_2003",
-              $"year_2004",
-              $"year_2005",
-              $"year_2006",
-              $"year_2007",
-              $"year_2008",
-              $"year_2009",
-              $"year_2010",
-              $"year_2011",
-              $"year_2012",
-              $"year_2013",
-              $"year_2014",
-              $"year_2015",
-              $"year_2016",
-              $"year_2017",
-              $"year_2018"
-            )
-            .agg(
-              sum("area") as "area",
-              sum("gain") as "gain",
-              sum("biomass") as "biomass",
-              sum($"biomass_per_ha" * $"area") / sum("area") as "biomass_per_ha",
-              sum("co2") as "co2",
-              sum("mangrove_biomass") as "mangrove_biomass",
-              sum($"mangrove_biomass_per_ha" * $"area") / sum("area") as "mangrove_biomass_per_ha",
-              sum("mangrove_co2") as "mangrove_co2",
-              sum("area_loss_2001") as "area_loss_2001",
-              sum("area_loss_2002") as "area_loss_2002",
-              sum("area_loss_2003") as "area_loss_2003",
-              sum("area_loss_2004") as "area_loss_2004",
-              sum("area_loss_2005") as "area_loss_2005",
-              sum("area_loss_2006") as "area_loss_2006",
-              sum("area_loss_2007") as "area_loss_2007",
-              sum("area_loss_2008") as "area_loss_2008",
-              sum("area_loss_2009") as "area_loss_2009",
-              sum("area_loss_2010") as "area_loss_2010",
-              sum("area_loss_2011") as "area_loss_2011",
-              sum("area_loss_2012") as "area_loss_2012",
-              sum("area_loss_2013") as "area_loss_2013",
-              sum("area_loss_2014") as "area_loss_2014",
-              sum("area_loss_2015") as "area_loss_2015",
-              sum("area_loss_2016") as "area_loss_2016",
-              sum("area_loss_2017") as "area_loss_2017",
-              sum("area_loss_2018") as "area_loss_2018",
-              sum("biomass_loss_2001") as "biomass_loss_2001",
-              sum("biomass_loss_2002") as "biomass_loss_2002",
-              sum("biomass_loss_2003") as "biomass_loss_2003",
-              sum("biomass_loss_2004") as "biomass_loss_2004",
-              sum("biomass_loss_2005") as "biomass_loss_2005",
-              sum("biomass_loss_2006") as "biomass_loss_2006",
-              sum("biomass_loss_2007") as "biomass_loss_2007",
-              sum("biomass_loss_2008") as "biomass_loss_2008",
-              sum("biomass_loss_2009") as "biomass_loss_2009",
-              sum("biomass_loss_2010") as "biomass_loss_2010",
-              sum("biomass_loss_2011") as "biomass_loss_2011",
-              sum("biomass_loss_2012") as "biomass_loss_2012",
-              sum("biomass_loss_2013") as "biomass_loss_2013",
-              sum("biomass_loss_2014") as "biomass_loss_2014",
-              sum("biomass_loss_2015") as "biomass_loss_2015",
-              sum("biomass_loss_2016") as "biomass_loss_2016",
-              sum("biomass_loss_2017") as "biomass_loss_2017",
-              sum("biomass_loss_2018") as "biomass_loss_2018",
-              sum("carbon_emissions_2001") as "carbon_emissions_2001",
-              sum("carbon_emissions_2002") as "carbon_emissions_2002",
-              sum("carbon_emissions_2003") as "carbon_emissions_2003",
-              sum("carbon_emissions_2004") as "carbon_emissions_2004",
-              sum("carbon_emissions_2005") as "carbon_emissions_2005",
-              sum("carbon_emissions_2006") as "carbon_emissions_2006",
-              sum("carbon_emissions_2007") as "carbon_emissions_2007",
-              sum("carbon_emissions_2008") as "carbon_emissions_2008",
-              sum("carbon_emissions_2009") as "carbon_emissions_2009",
-              sum("carbon_emissions_2010") as "carbon_emissions_2010",
-              sum("carbon_emissions_2011") as "carbon_emissions_2011",
-              sum("carbon_emissions_2012") as "carbon_emissions_2012",
-              sum("carbon_emissions_2013") as "carbon_emissions_2013",
-              sum("carbon_emissions_2014") as "carbon_emissions_2014",
-              sum("carbon_emissions_2015") as "carbon_emissions_2015",
-              sum("carbon_emissions_2016") as "carbon_emissions_2016",
-              sum("carbon_emissions_2017") as "carbon_emissions_2017",
-              sum("carbon_emissions_2018") as "carbon_emissions_2018",
-              sum("mangrove_biomass_loss_2001") as "mangrove_biomass_loss_2001",
-              sum("mangrove_biomass_loss_2002") as "mangrove_biomass_loss_2002",
-              sum("mangrove_biomass_loss_2003") as "mangrove_biomass_loss_2003",
-              sum("mangrove_biomass_loss_2004") as "mangrove_biomass_loss_2004",
-              sum("mangrove_biomass_loss_2005") as "mangrove_biomass_loss_2005",
-              sum("mangrove_biomass_loss_2006") as "mangrove_biomass_loss_2006",
-              sum("mangrove_biomass_loss_2007") as "mangrove_biomass_loss_2007",
-              sum("mangrove_biomass_loss_2008") as "mangrove_biomass_loss_2008",
-              sum("mangrove_biomass_loss_2009") as "mangrove_biomass_loss_2009",
-              sum("mangrove_biomass_loss_2010") as "mangrove_biomass_loss_2010",
-              sum("mangrove_biomass_loss_2011") as "mangrove_biomass_loss_2011",
-              sum("mangrove_biomass_loss_2012") as "mangrove_biomass_loss_2012",
-              sum("mangrove_biomass_loss_2013") as "mangrove_biomass_loss_2013",
-              sum("mangrove_biomass_loss_2014") as "mangrove_biomass_loss_2014",
-              sum("mangrove_biomass_loss_2015") as "mangrove_biomass_loss_2015",
-              sum("mangrove_biomass_loss_2016") as "mangrove_biomass_loss_2016",
-              sum("mangrove_biomass_loss_2017") as "mangrove_biomass_loss_2017",
-              sum("mangrove_biomass_loss_2018") as "mangrove_biomass_loss_2018",
-              sum("mangrove_carbon_emissions_2001") as "mangrove_carbon_emissions_2001",
-              sum("mangrove_carbon_emissions_2002") as "mangrove_carbon_emissions_2002",
-              sum("mangrove_carbon_emissions_2003") as "mangrove_carbon_emissions_2003",
-              sum("mangrove_carbon_emissions_2004") as "mangrove_carbon_emissions_2004",
-              sum("mangrove_carbon_emissions_2005") as "mangrove_carbon_emissions_2005",
-              sum("mangrove_carbon_emissions_2006") as "mangrove_carbon_emissions_2006",
-              sum("mangrove_carbon_emissions_2007") as "mangrove_carbon_emissions_2007",
-              sum("mangrove_carbon_emissions_2008") as "mangrove_carbon_emissions_2008",
-              sum("mangrove_carbon_emissions_2009") as "mangrove_carbon_emissions_2009",
-              sum("mangrove_carbon_emissions_2010") as "mangrove_carbon_emissions_2010",
-              sum("mangrove_carbon_emissions_2011") as "mangrove_carbon_emissions_2011",
-              sum("mangrove_carbon_emissions_2012") as "mangrove_carbon_emissions_2012",
-              sum("mangrove_carbon_emissions_2013") as "mangrove_carbon_emissions_2013",
-              sum("mangrove_carbon_emissions_2014") as "mangrove_carbon_emissions_2014",
-              sum("mangrove_carbon_emissions_2015") as "mangrove_carbon_emissions_2015",
-              sum("mangrove_carbon_emissions_2016") as "mangrove_carbon_emissions_2016",
-              sum("mangrove_carbon_emissions_2017") as "mangrove_carbon_emissions_2017",
-              sum("mangrove_carbon_emissions_2018") as "mangrove_carbon_emissions_2018"
-            )
-            .join(
-              masterDF,
-              $"feature_id" <=> $"m_feature_id"
-                && $"layers" <=> $"m_layers" && $"threshold" <=> $"m_threshold",
-              "right_outer"
-            )
-            .na
-            .fill(
-              0.0,
-              Seq(
-                "area",
-                "gain",
-                "biomass",
-                "biomass_per_ha",
-                "co2",
-                "mangrove_biomass",
-                "mangrove_biomass_per_ha",
-                "mangrove_co2",
-                "area_loss_2001",
-                "area_loss_2002",
-                "area_loss_2003",
-                "area_loss_2004",
-                "area_loss_2005",
-                "area_loss_2006",
-                "area_loss_2007",
-                "area_loss_2008",
-                "area_loss_2009",
-                "area_loss_2010",
-                "area_loss_2011",
-                "area_loss_2012",
-                "area_loss_2013",
-                "area_loss_2014",
-                "area_loss_2015",
-                "area_loss_2016",
-                "area_loss_2017",
-                "area_loss_2018",
-                "biomass_loss_2001",
-                "biomass_loss_2002",
-                "biomass_loss_2003",
-                "biomass_loss_2004",
-                "biomass_loss_2005",
-                "biomass_loss_2006",
-                "biomass_loss_2007",
-                "biomass_loss_2008",
-                "biomass_loss_2009",
-                "biomass_loss_2010",
-                "biomass_loss_2011",
-                "biomass_loss_2012",
-                "biomass_loss_2013",
-                "biomass_loss_2014",
-                "biomass_loss_2015",
-                "biomass_loss_2016",
-                "biomass_loss_2017",
-                "biomass_loss_2018",
-                "carbon_emissions_2001",
-                "carbon_emissions_2002",
-                "carbon_emissions_2003",
-                "carbon_emissions_2004",
-                "carbon_emissions_2005",
-                "carbon_emissions_2006",
-                "carbon_emissions_2007",
-                "carbon_emissions_2008",
-                "carbon_emissions_2009",
-                "carbon_emissions_2010",
-                "carbon_emissions_2011",
-                "carbon_emissions_2012",
-                "carbon_emissions_2013",
-                "carbon_emissions_2014",
-                "carbon_emissions_2015",
-                "carbon_emissions_2016",
-                "carbon_emissions_2017",
-                "carbon_emissions_2018",
-                "mangrove_biomass_loss_2001",
-                "mangrove_biomass_loss_2002",
-                "mangrove_biomass_loss_2003",
-                "mangrove_biomass_loss_2004",
-                "mangrove_biomass_loss_2005",
-                "mangrove_biomass_loss_2006",
-                "mangrove_biomass_loss_2007",
-                "mangrove_biomass_loss_2008",
-                "mangrove_biomass_loss_2009",
-                "mangrove_biomass_loss_2010",
-                "mangrove_biomass_loss_2011",
-                "mangrove_biomass_loss_2012",
-                "mangrove_biomass_loss_2013",
-                "mangrove_biomass_loss_2014",
-                "mangrove_biomass_loss_2015",
-                "mangrove_biomass_loss_2016",
-                "mangrove_biomass_loss_2017",
-                "mangrove_biomass_loss_2018",
-                "mangrove_carbon_emissions_2001",
-                "mangrove_carbon_emissions_2002",
-                "mangrove_carbon_emissions_2003",
-                "mangrove_carbon_emissions_2004",
-                "mangrove_carbon_emissions_2005",
-                "mangrove_carbon_emissions_2006",
-                "mangrove_carbon_emissions_2007",
-                "mangrove_carbon_emissions_2008",
-                "mangrove_carbon_emissions_2009",
-                "mangrove_carbon_emissions_2010",
-                "mangrove_carbon_emissions_2011",
-                "mangrove_carbon_emissions_2012",
-                "mangrove_carbon_emissions_2013",
-                "mangrove_carbon_emissions_2014",
-                "mangrove_carbon_emissions_2015",
-                "mangrove_carbon_emissions_2016",
-                "mangrove_carbon_emissions_2017",
-                "mangrove_carbon_emissions_2018"
-              )
-            )
-            .na
-            .fill(2001, Seq("year_2001"))
-            .na
-            .fill(2002, Seq("year_2002"))
-            .na
-            .fill(2003, Seq("year_2003"))
-            .na
-            .fill(2004, Seq("year_2004"))
-            .na
-            .fill(2005, Seq("year_2005"))
-            .na
-            .fill(2006, Seq("year_2006"))
-            .na
-            .fill(2007, Seq("year_2007"))
-            .na
-            .fill(2008, Seq("year_2008"))
-            .na
-            .fill(2009, Seq("year_2009"))
-            .na
-            .fill(2010, Seq("year_2010"))
-            .na
-            .fill(2011, Seq("year_2011"))
-            .na
-            .fill(2012, Seq("year_2012"))
-            .na
-            .fill(2013, Seq("year_2013"))
-            .na
-            .fill(2014, Seq("year_2014"))
-            .na
-            .fill(2015, Seq("year_2015"))
-            .na
-            .fill(2016, Seq("year_2016"))
-            .na
-            .fill(2017, Seq("year_2017"))
-            .na
-            .fill(2018, Seq("year_2018"))
-            .select(
-              $"m_feature_id" as "feature_id",
-              $"m_threshold" as "threshold",
-              $"m_layers" as "layers",
-              $"totalarea",
-              windowSum("area") as "extent2000",
-              windowSum("gain") as "total_gain",
-              windowSum("biomass") as "total_biomass",
-              windowSum($"biomass_per_ha" * $"area") / windowSum("area") as "avg_biomass_per_ha",
-              windowSum("co2") as "total_co2",
-              windowSum("mangrove_biomass") as "total_mangrove_biomass",
-              windowSum("mangrove_co2") as "total_mangrove_co2",
-              windowSum($"mangrove_biomass_per_ha" * $"area") / windowSum("area") as "avg_mangrove_biomass_per_ha",
-              $"year_2001",
-              $"year_2002",
-              $"year_2003",
-              $"year_2004",
-              $"year_2005",
-              $"year_2006",
-              $"year_2007",
-              $"year_2008",
-              $"year_2009",
-              $"year_2010",
-              $"year_2011",
-              $"year_2012",
-              $"year_2013",
-              $"year_2014",
-              $"year_2015",
-              $"year_2016",
-              $"year_2017",
-              $"year_2018",
-              windowSum("area_loss_2001") as "area_loss_2001",
-              windowSum("area_loss_2002") as "area_loss_2002",
-              windowSum("area_loss_2003") as "area_loss_2003",
-              windowSum("area_loss_2004") as "area_loss_2004",
-              windowSum("area_loss_2005") as "area_loss_2005",
-              windowSum("area_loss_2006") as "area_loss_2006",
-              windowSum("area_loss_2007") as "area_loss_2007",
-              windowSum("area_loss_2008") as "area_loss_2008",
-              windowSum("area_loss_2009") as "area_loss_2009",
-              windowSum("area_loss_2010") as "area_loss_2010",
-              windowSum("area_loss_2011") as "area_loss_2011",
-              windowSum("area_loss_2012") as "area_loss_2012",
-              windowSum("area_loss_2013") as "area_loss_2013",
-              windowSum("area_loss_2014") as "area_loss_2014",
-              windowSum("area_loss_2015") as "area_loss_2015",
-              windowSum("area_loss_2016") as "area_loss_2016",
-              windowSum("area_loss_2017") as "area_loss_2017",
-              windowSum("area_loss_2018") as "area_loss_2018",
-              windowSum("biomass_loss_2001") as "biomass_loss_2001",
-              windowSum("biomass_loss_2002") as "biomass_loss_2002",
-              windowSum("biomass_loss_2003") as "biomass_loss_2003",
-              windowSum("biomass_loss_2004") as "biomass_loss_2004",
-              windowSum("biomass_loss_2005") as "biomass_loss_2005",
-              windowSum("biomass_loss_2006") as "biomass_loss_2006",
-              windowSum("biomass_loss_2007") as "biomass_loss_2007",
-              windowSum("biomass_loss_2008") as "biomass_loss_2008",
-              windowSum("biomass_loss_2009") as "biomass_loss_2009",
-              windowSum("biomass_loss_2010") as "biomass_loss_2010",
-              windowSum("biomass_loss_2011") as "biomass_loss_2011",
-              windowSum("biomass_loss_2012") as "biomass_loss_2012",
-              windowSum("biomass_loss_2013") as "biomass_loss_2013",
-              windowSum("biomass_loss_2014") as "biomass_loss_2014",
-              windowSum("biomass_loss_2015") as "biomass_loss_2015",
-              windowSum("biomass_loss_2016") as "biomass_loss_2016",
-              windowSum("biomass_loss_2017") as "biomass_loss_2017",
-              windowSum("biomass_loss_2018") as "biomass_loss_2018",
-              windowSum("carbon_emissions_2001") as "carbon_emissions_2001",
-              windowSum("carbon_emissions_2002") as "carbon_emissions_2002",
-              windowSum("carbon_emissions_2003") as "carbon_emissions_2003",
-              windowSum("carbon_emissions_2004") as "carbon_emissions_2004",
-              windowSum("carbon_emissions_2005") as "carbon_emissions_2005",
-              windowSum("carbon_emissions_2006") as "carbon_emissions_2006",
-              windowSum("carbon_emissions_2007") as "carbon_emissions_2007",
-              windowSum("carbon_emissions_2008") as "carbon_emissions_2008",
-              windowSum("carbon_emissions_2009") as "carbon_emissions_2009",
-              windowSum("carbon_emissions_2010") as "carbon_emissions_2010",
-              windowSum("carbon_emissions_2011") as "carbon_emissions_2011",
-              windowSum("carbon_emissions_2012") as "carbon_emissions_2012",
-              windowSum("carbon_emissions_2013") as "carbon_emissions_2013",
-              windowSum("carbon_emissions_2014") as "carbon_emissions_2014",
-              windowSum("carbon_emissions_2015") as "carbon_emissions_2015",
-              windowSum("carbon_emissions_2016") as "carbon_emissions_2016",
-              windowSum("carbon_emissions_2017") as "carbon_emissions_2017",
-              windowSum("carbon_emissions_2018") as "carbon_emissions_2018",
-              windowSum("mangrove_biomass_loss_2001") as "mangrove_biomass_loss_2001",
-              windowSum("mangrove_biomass_loss_2002") as "mangrove_biomass_loss_2002",
-              windowSum("mangrove_biomass_loss_2003") as "mangrove_biomass_loss_2003",
-              windowSum("mangrove_biomass_loss_2004") as "mangrove_biomass_loss_2004",
-              windowSum("mangrove_biomass_loss_2005") as "mangrove_biomass_loss_2005",
-              windowSum("mangrove_biomass_loss_2006") as "mangrove_biomass_loss_2006",
-              windowSum("mangrove_biomass_loss_2007") as "mangrove_biomass_loss_2007",
-              windowSum("mangrove_biomass_loss_2008") as "mangrove_biomass_loss_2008",
-              windowSum("mangrove_biomass_loss_2009") as "mangrove_biomass_loss_2009",
-              windowSum("mangrove_biomass_loss_2010") as "mangrove_biomass_loss_2010",
-              windowSum("mangrove_biomass_loss_2011") as "mangrove_biomass_loss_2011",
-              windowSum("mangrove_biomass_loss_2012") as "mangrove_biomass_loss_2012",
-              windowSum("mangrove_biomass_loss_2013") as "mangrove_biomass_loss_2013",
-              windowSum("mangrove_biomass_loss_2014") as "mangrove_biomass_loss_2014",
-              windowSum("mangrove_biomass_loss_2015") as "mangrove_biomass_loss_2015",
-              windowSum("mangrove_biomass_loss_2016") as "mangrove_biomass_loss_2016",
-              windowSum("mangrove_biomass_loss_2017") as "mangrove_biomass_loss_2017",
-              windowSum("mangrove_biomass_loss_2018") as "mangrove_biomass_loss_2018",
-              windowSum("mangrove_carbon_emissions_2001") as "mangrove_carbon_emissions_2001",
-              windowSum("mangrove_carbon_emissions_2002") as "mangrove_carbon_emissions_2002",
-              windowSum("mangrove_carbon_emissions_2003") as "mangrove_carbon_emissions_2003",
-              windowSum("mangrove_carbon_emissions_2004") as "mangrove_carbon_emissions_2004",
-              windowSum("mangrove_carbon_emissions_2005") as "mangrove_carbon_emissions_2005",
-              windowSum("mangrove_carbon_emissions_2006") as "mangrove_carbon_emissions_2006",
-              windowSum("mangrove_carbon_emissions_2007") as "mangrove_carbon_emissions_2007",
-              windowSum("mangrove_carbon_emissions_2008") as "mangrove_carbon_emissions_2008",
-              windowSum("mangrove_carbon_emissions_2009") as "mangrove_carbon_emissions_2009",
-              windowSum("mangrove_carbon_emissions_2010") as "mangrove_carbon_emissions_2010",
-              windowSum("mangrove_carbon_emissions_2011") as "mangrove_carbon_emissions_2011",
-              windowSum("mangrove_carbon_emissions_2012") as "mangrove_carbon_emissions_2012",
-              windowSum("mangrove_carbon_emissions_2013") as "mangrove_carbon_emissions_2013",
-              windowSum("mangrove_carbon_emissions_2014") as "mangrove_carbon_emissions_2014",
-              windowSum("mangrove_carbon_emissions_2015") as "mangrove_carbon_emissions_2015",
-              windowSum("mangrove_carbon_emissions_2016") as "mangrove_carbon_emissions_2016",
-              windowSum("mangrove_carbon_emissions_2017") as "mangrove_carbon_emissions_2017",
-              windowSum("mangrove_carbon_emissions_2018") as "mangrove_carbon_emissions_2018"
-            )
-
+            .transform(AnnualLossDF.unpackYearData)
+            .transform(AnnualLossDF.sumArea)
+            .transform(AnnualLossDF.joinMaster(masterDF))
+            .transform(AnnualLossDF.aggregateByThreshold)
 
           val adm2DF = annualLossDF
-            .join(
-              extent2010DF,
-              $"m_feature_id" <=> $"feature_id" &&
-                $"m_layers" <=> $"layers" &&
-                $"m_threshold" <=> $"threshold",
-              "left_outer"
-            )
-            .select(
-              $"feature_id.iso" as "iso",
-              $"feature_id.adm1" as "adm1",
-              $"feature_id.adm2" as "adm2",
-              $"threshold",
-              $"layers.drivers" as "tcs",
-              $"layers.globalLandCover" as "global_land_cover",
-              $"layers.primaryForest" as "primary_forest",
-              $"layers.idnPrimaryForest" as "idn_primary_forest",
-              $"layers.erosion" as "erosion",
-              $"layers.biodiversitySignificance" as "biodiversity_significance",
-              $"layers.biodiversityIntactness" as "biodiversity_intactness",
-              $"layers.wdpa" as "wdpa",
-              $"layers.aze" as "aze",
-              $"layers.plantations" as "plantations",
-              $"layers.riverBasins" as "river_basin",
-              $"layers.ecozones" as "ecozone",
-              $"layers.urbanWatersheds" as "urban_watershed",
-              $"layers.mangroves1996" as "mangroves_1996",
-              $"layers.mangroves2016" as "mangroves_2016",
-              $"layers.waterStress" as "water_stress",
-              $"layers.intactForestLandscapes" as "ifl",
-              $"layers.endemicBirdAreas" as "endemic_bird_area",
-              $"layers.tigerLandscapes" as "tiger_cl",
-              $"layers.landmark" as "landmark",
-              $"layers.landRights" as "land_right",
-              $"layers.keyBiodiversityAreas" as "kba",
-              $"layers.mining" as "mining",
-              $"layers.rspo" as "rspo",
-              $"layers.peatlands" as "idn_mys_peatlands",
-              $"layers.oilPalm" as "oil_palm",
-              $"layers.idnForestMoratorium" as "idn_forest_moratorium",
-              $"layers.idnLandCover" as "idn_land_cover",
-              $"layers.mexProtectedAreas" as "mex_protected_areas",
-              $"layers.mexPaymentForEcosystemServices" as "mex_pes",
-              $"layers.mexForestZoning" as "mex_forest_zoning",
-              $"layers.perProductionForest" as "per_production_forest",
-              $"layers.perProtectedAreas" as "per_protected_area",
-              $"layers.perForestConcessions" as "per_forest_concession",
-              $"layers.braBiomes" as "bra_biomes",
-              $"layers.woodFiber" as "wood_fiber",
-              $"layers.resourceRights" as "resource_right",
-              $"layers.logging" as "managed_forests",
-              $"layers.oilGas" as "oil_gas",
-              $"totalarea",
-              $"extent2000",
-              $"extent2010",
-              $"total_gain",
-              $"total_biomass",
-              $"avg_biomass_per_ha",
-              $"total_co2",
-              $"total_mangrove_biomass",
-              $"avg_mangrove_biomass_per_ha",
-              $"total_mangrove_co2",
-              $"year_2001",
-              $"area_loss_2001",
-              $"biomass_loss_2001",
-              $"carbon_emissions_2001",
-              $"mangrove_biomass_loss_2001",
-              $"mangrove_carbon_emissions_2001",
-              $"year_2002",
-              $"area_loss_2002",
-              $"biomass_loss_2002",
-              $"carbon_emissions_2002",
-              $"mangrove_biomass_loss_2002",
-              $"mangrove_carbon_emissions_2002",
-              $"year_2003",
-              $"area_loss_2003",
-              $"biomass_loss_2003",
-              $"carbon_emissions_2003",
-              $"mangrove_biomass_loss_2003",
-              $"mangrove_carbon_emissions_2003",
-              $"year_2004",
-              $"area_loss_2004",
-              $"biomass_loss_2004",
-              $"carbon_emissions_2004",
-              $"mangrove_biomass_loss_2004",
-              $"mangrove_carbon_emissions_2004",
-              $"year_2005",
-              $"area_loss_2005",
-              $"biomass_loss_2005",
-              $"carbon_emissions_2005",
-              $"mangrove_biomass_loss_2005",
-              $"mangrove_carbon_emissions_2005",
-              $"year_2006",
-              $"area_loss_2006",
-              $"biomass_loss_2006",
-              $"carbon_emissions_2006",
-              $"mangrove_biomass_loss_2006",
-              $"mangrove_carbon_emissions_2006",
-              $"year_2007",
-              $"area_loss_2007",
-              $"biomass_loss_2007",
-              $"carbon_emissions_2007",
-              $"mangrove_biomass_loss_2007",
-              $"mangrove_carbon_emissions_2007",
-              $"year_2008",
-              $"area_loss_2008",
-              $"biomass_loss_2008",
-              $"carbon_emissions_2008",
-              $"mangrove_biomass_loss_2008",
-              $"mangrove_carbon_emissions_2008",
-              $"year_2009",
-              $"area_loss_2009",
-              $"biomass_loss_2009",
-              $"carbon_emissions_2009",
-              $"mangrove_biomass_loss_2009",
-              $"mangrove_carbon_emissions_2009",
-              $"year_2010",
-              $"area_loss_2010",
-              $"biomass_loss_2010",
-              $"carbon_emissions_2010",
-              $"mangrove_biomass_loss_2010",
-              $"mangrove_carbon_emissions_2010",
-              $"year_2011",
-              $"area_loss_2011",
-              $"biomass_loss_2011",
-              $"carbon_emissions_2011",
-              $"mangrove_biomass_loss_2011",
-              $"mangrove_carbon_emissions_2011",
-              $"year_2012",
-              $"area_loss_2012",
-              $"biomass_loss_2012",
-              $"carbon_emissions_2012",
-              $"mangrove_biomass_loss_2012",
-              $"mangrove_carbon_emissions_2012",
-              $"year_2013",
-              $"area_loss_2013",
-              $"biomass_loss_2013",
-              $"carbon_emissions_2013",
-              $"mangrove_biomass_loss_2013",
-              $"mangrove_carbon_emissions_2013",
-              $"year_2014",
-              $"area_loss_2014",
-              $"biomass_loss_2014",
-              $"carbon_emissions_2014",
-              $"mangrove_biomass_loss_2014",
-              $"mangrove_carbon_emissions_2014",
-              $"year_2015",
-              $"area_loss_2015",
-              $"biomass_loss_2015",
-              $"carbon_emissions_2015",
-              $"mangrove_biomass_loss_2015",
-              $"mangrove_carbon_emissions_2015",
-              $"year_2016",
-              $"area_loss_2016",
-              $"biomass_loss_2016",
-              $"carbon_emissions_2016",
-              $"mangrove_biomass_loss_2016",
-              $"mangrove_carbon_emissions_2016",
-              $"year_2017",
-              $"area_loss_2017",
-              $"biomass_loss_2017",
-              $"carbon_emissions_2017",
-              $"mangrove_biomass_loss_2017",
-              $"mangrove_carbon_emissions_2017",
-              $"year_2018",
-              $"area_loss_2018",
-              $"biomass_loss_2018",
-              $"carbon_emissions_2018",
-              $"mangrove_biomass_loss_2018",
-              $"mangrove_carbon_emissions_2018"
-            )
+            .transform(Adm2DF.joinExtent2010(extent2010DF))
+            .transform(Adm2DF.unpackFeautureIDLayers)
 
           adm2DF.repartition($"iso")
           adm2DF.cache()
@@ -1106,1176 +243,79 @@ object TreeLossSummaryMain
           )
 
           val adm2SummaryDF = adm2DF
-            .groupBy($"iso", $"adm1", $"adm2", $"threshold")
-            .agg(
-              round(sum("totalarea")) as "area_ha",
-              round(sum("extent2000")) as "extent2000_ha",
-              round(sum("extent2010")) as "extent2010_ha",
-              round(sum("total_gain")) as "gain_2000_2012_ha",
-              round(sum("total_biomass")) as "biomass_Mt",
-              round(sum($"avg_biomass_per_ha" * $"totalarea") / sum("totalarea")) as "avg_biomass_per_ha_Mt",
-              round(sum("total_co2")) as "carbon_Mt",
-              round(sum("total_mangrove_biomass")) as "mangrove_biomass_Mt",
-              round(sum($"avg_mangrove_biomass_per_ha" * $"totalarea") / sum("totalarea")) as "avg_mangrove_biomass_per_ha_Mt",
-              round(sum("total_mangrove_co2")) as "mangrove_carbon_Mt",
-              round(sum("area_loss_2001")) as "tc_loss_ha_2001",
-              round(sum("area_loss_2002")) as "tc_loss_ha_2002",
-              round(sum("area_loss_2003")) as "tc_loss_ha_2003",
-              round(sum("area_loss_2004")) as "tc_loss_ha_2004",
-              round(sum("area_loss_2005")) as "tc_loss_ha_2005",
-              round(sum("area_loss_2006")) as "tc_loss_ha_2006",
-              round(sum("area_loss_2007")) as "tc_loss_ha_2007",
-              round(sum("area_loss_2008")) as "tc_loss_ha_2008",
-              round(sum("area_loss_2009")) as "tc_loss_ha_2009",
-              round(sum("area_loss_2010")) as "tc_loss_ha_2010",
-              round(sum("area_loss_2011")) as "tc_loss_ha_2011",
-              round(sum("area_loss_2012")) as "tc_loss_ha_2012",
-              round(sum("area_loss_2013")) as "tc_loss_ha_2013",
-              round(sum("area_loss_2014")) as "tc_loss_ha_2014",
-              round(sum("area_loss_2015")) as "tc_loss_ha_2015",
-              round(sum("area_loss_2016")) as "tc_loss_ha_2016",
-              round(sum("area_loss_2017")) as "tc_loss_ha_2017",
-              round(sum("area_loss_2018")) as "tc_loss_ha_2018",
-              round(sum("biomass_loss_2001")) as "biomass_loss_Mt_2001",
-              round(sum("biomass_loss_2002")) as "biomass_loss_Mt_2002",
-              round(sum("biomass_loss_2003")) as "biomass_loss_Mt_2003",
-              round(sum("biomass_loss_2004")) as "biomass_loss_Mt_2004",
-              round(sum("biomass_loss_2005")) as "biomass_loss_Mt_2005",
-              round(sum("biomass_loss_2006")) as "biomass_loss_Mt_2006",
-              round(sum("biomass_loss_2007")) as "biomass_loss_Mt_2007",
-              round(sum("biomass_loss_2008")) as "biomass_loss_Mt_2008",
-              round(sum("biomass_loss_2009")) as "biomass_loss_Mt_2009",
-              round(sum("biomass_loss_2010")) as "biomass_loss_Mt_2010",
-              round(sum("biomass_loss_2011")) as "biomass_loss_Mt_2011",
-              round(sum("biomass_loss_2012")) as "biomass_loss_Mt_2012",
-              round(sum("biomass_loss_2013")) as "biomass_loss_Mt_2013",
-              round(sum("biomass_loss_2014")) as "biomass_loss_Mt_2014",
-              round(sum("biomass_loss_2015")) as "biomass_loss_Mt_2015",
-              round(sum("biomass_loss_2016")) as "biomass_loss_Mt_2016",
-              round(sum("biomass_loss_2017")) as "biomass_loss_Mt_2017",
-              round(sum("biomass_loss_2018")) as "biomass_loss_Mt_2018",
-              round(sum("carbon_emissions_2001")) as "carbon_emissions_Mt_2001",
-              round(sum("carbon_emissions_2002")) as "carbon_emissions_Mt_2002",
-              round(sum("carbon_emissions_2003")) as "carbon_emissions_Mt_2003",
-              round(sum("carbon_emissions_2004")) as "carbon_emissions_Mt_2004",
-              round(sum("carbon_emissions_2005")) as "carbon_emissions_Mt_2005",
-              round(sum("carbon_emissions_2006")) as "carbon_emissions_Mt_2006",
-              round(sum("carbon_emissions_2007")) as "carbon_emissions_Mt_2007",
-              round(sum("carbon_emissions_2008")) as "carbon_emissions_Mt_2008",
-              round(sum("carbon_emissions_2009")) as "carbon_emissions_Mt_2009",
-              round(sum("carbon_emissions_2010")) as "carbon_emissions_Mt_2010",
-              round(sum("carbon_emissions_2011")) as "carbon_emissions_Mt_2011",
-              round(sum("carbon_emissions_2012")) as "carbon_emissions_Mt_2012",
-              round(sum("carbon_emissions_2013")) as "carbon_emissions_Mt_2013",
-              round(sum("carbon_emissions_2014")) as "carbon_emissions_Mt_2014",
-              round(sum("carbon_emissions_2015")) as "carbon_emissions_Mt_2015",
-              round(sum("carbon_emissions_2016")) as "carbon_emissions_Mt_2016",
-              round(sum("carbon_emissions_2017")) as "carbon_emissions_Mt_2017",
-              round(sum("carbon_emissions_2018")) as "carbon_emissions_Mt_2018"
-            )
+            .filter($"threshold" > 0)
+            .transform(Adm2SummaryDF.sumArea)
 
-          adm2SummaryDF.write
+          adm2SummaryDF
+            .repartition($"iso")
+            .orderBy($"iso", $"adm1", $"adm2", $"threshold")
+            .write
             .options(csvOptions)
             .csv(path = runOutputUrl + "/summary/adm2")
 
-          val adm1SummaryDF = adm2SummaryDF
-            .groupBy($"iso", $"adm1", $"threshold")
-            .agg(
-              round(sum("area_ha")) as "area_ha",
-              round(sum("extent2000_ha")) as "extent2000_ha",
-              round(sum("extent2010_ha")) as "extent2010_ha",
-              round(sum("gain_2000_2012_ha")) as "gain_2000_2012_ha",
-              round(sum("biomass_Mt")) as "biomass_Mt",
-              round(sum($"avg_biomass_per_ha_Mt" * $"area_ha") / sum("area_ha")) as "avg_biomass_per_ha_Mt",
-              round(sum("carbon_Mt")) as "carbon_Mt",
-              round(sum("mangrove_biomass_Mt")) as "mangrove_biomass_Mt",
-              round(sum($"avg_mangrove_biomass_per_ha_Mt" * $"area_ha") / sum("area_ha")) as "avg_mangrove_biomass_per_ha_Mt",
-              round(sum("mangrove_carbon_Mt")) as "mangrove_carbon_Mt",
-              round(sum("tc_loss_ha_2001")) as "tc_loss_ha_2001",
-              round(sum("tc_loss_ha_2002")) as "tc_loss_ha_2002",
-              round(sum("tc_loss_ha_2003")) as "tc_loss_ha_2003",
-              round(sum("tc_loss_ha_2004")) as "tc_loss_ha_2004",
-              round(sum("tc_loss_ha_2005")) as "tc_loss_ha_2005",
-              round(sum("tc_loss_ha_2006")) as "tc_loss_ha_2006",
-              round(sum("tc_loss_ha_2007")) as "tc_loss_ha_2007",
-              round(sum("tc_loss_ha_2008")) as "tc_loss_ha_2008",
-              round(sum("tc_loss_ha_2009")) as "tc_loss_ha_2009",
-              round(sum("tc_loss_ha_2010")) as "tc_loss_ha_2010",
-              round(sum("tc_loss_ha_2011")) as "tc_loss_ha_2011",
-              round(sum("tc_loss_ha_2012")) as "tc_loss_ha_2012",
-              round(sum("tc_loss_ha_2013")) as "tc_loss_ha_2013",
-              round(sum("tc_loss_ha_2014")) as "tc_loss_ha_2014",
-              round(sum("tc_loss_ha_2015")) as "tc_loss_ha_2015",
-              round(sum("tc_loss_ha_2016")) as "tc_loss_ha_2016",
-              round(sum("tc_loss_ha_2017")) as "tc_loss_ha_2017",
-              round(sum("tc_loss_ha_2018")) as "tc_loss_ha_2018",
-              round(sum("biomass_loss_Mt_2001")) as "biomass_loss_Mt_2001",
-              round(sum("biomass_loss_Mt_2002")) as "biomass_loss_Mt_2002",
-              round(sum("biomass_loss_Mt_2003")) as "biomass_loss_Mt_2003",
-              round(sum("biomass_loss_Mt_2004")) as "biomass_loss_Mt_2004",
-              round(sum("biomass_loss_Mt_2005")) as "biomass_loss_Mt_2005",
-              round(sum("biomass_loss_Mt_2006")) as "biomass_loss_Mt_2006",
-              round(sum("biomass_loss_Mt_2007")) as "biomass_loss_Mt_2007",
-              round(sum("biomass_loss_Mt_2008")) as "biomass_loss_Mt_2008",
-              round(sum("biomass_loss_Mt_2009")) as "biomass_loss_Mt_2009",
-              round(sum("biomass_loss_Mt_2010")) as "biomass_loss_Mt_2010",
-              round(sum("biomass_loss_Mt_2011")) as "biomass_loss_Mt_2011",
-              round(sum("biomass_loss_Mt_2012")) as "biomass_loss_Mt_2012",
-              round(sum("biomass_loss_Mt_2013")) as "biomass_loss_Mt_2013",
-              round(sum("biomass_loss_Mt_2014")) as "biomass_loss_Mt_2014",
-              round(sum("biomass_loss_Mt_2015")) as "biomass_loss_Mt_2015",
-              round(sum("biomass_loss_Mt_2016")) as "biomass_loss_Mt_2016",
-              round(sum("biomass_loss_Mt_2017")) as "biomass_loss_Mt_2017",
-              round(sum("biomass_loss_Mt_2018")) as "biomass_loss_Mt_2018",
-              round(sum("carbon_emissions_Mt_2001")) as "carbon_emissions_Mt_2001",
-              round(sum("carbon_emissions_Mt_2002")) as "carbon_emissions_Mt_2002",
-              round(sum("carbon_emissions_Mt_2003")) as "carbon_emissions_Mt_2003",
-              round(sum("carbon_emissions_Mt_2004")) as "carbon_emissions_Mt_2004",
-              round(sum("carbon_emissions_Mt_2005")) as "carbon_emissions_Mt_2005",
-              round(sum("carbon_emissions_Mt_2006")) as "carbon_emissions_Mt_2006",
-              round(sum("carbon_emissions_Mt_2007")) as "carbon_emissions_Mt_2007",
-              round(sum("carbon_emissions_Mt_2008")) as "carbon_emissions_Mt_2008",
-              round(sum("carbon_emissions_Mt_2009")) as "carbon_emissions_Mt_2009",
-              round(sum("carbon_emissions_Mt_2010")) as "carbon_emissions_Mt_2010",
-              round(sum("carbon_emissions_Mt_2011")) as "carbon_emissions_Mt_2011",
-              round(sum("carbon_emissions_Mt_2012")) as "carbon_emissions_Mt_2012",
-              round(sum("carbon_emissions_Mt_2013")) as "carbon_emissions_Mt_2013",
-              round(sum("carbon_emissions_Mt_2014")) as "carbon_emissions_Mt_2014",
-              round(sum("carbon_emissions_Mt_2015")) as "carbon_emissions_Mt_2015",
-              round(sum("carbon_emissions_Mt_2016")) as "carbon_emissions_Mt_2016",
-              round(sum("carbon_emissions_Mt_2017")) as "carbon_emissions_Mt_2017",
-              round(sum("carbon_emissions_Mt_2018")) as "carbon_emissions_Mt_2018"
-            )
+          val adm1SummaryDF = adm2SummaryDF.transform(Adm1SummaryDF.sumArea)
 
-          adm1SummaryDF.write
+          adm1SummaryDF
+            .repartition($"iso")
+            .orderBy($"iso", $"adm1", $"threshold")
+            .write
             .options(csvOptions)
             .csv(path = runOutputUrl + "/summary/adm1")
 
-          val isoSummaryDF = adm1SummaryDF
-            .groupBy($"iso", $"threshold")
-            .agg(
-              round(sum("area_ha")) as "area_ha",
-              round(sum("extent2000_ha")) as "extent2000_ha",
-              round(sum("extent2010_ha")) as "extent2010_ha",
-              round(sum("gain_2000_2012_ha")) as "gain_2000_2012_ha",
-              round(sum("biomass_Mt")) as "biomass_Mt",
-              round(sum($"avg_biomass_per_ha_Mt" * $"area_ha") / sum("area_ha")) as "avg_biomass_per_ha_Mt",
-              round(sum("carbon_Mt")) as "carbon_Mt",
-              round(sum("mangrove_biomass_Mt")) as "mangrove_biomass_Mt",
-              round(sum($"avg_mangrove_biomass_per_ha_Mt" * $"area_ha") / sum("area_ha")) as "avg_mangrove_biomass_per_ha_Mt",
-              round(sum("mangrove_carbon_Mt")) as "mangrove_carbon_Mt",
-              round(sum("tc_loss_ha_2001")) as "tc_loss_ha_2001",
-              round(sum("tc_loss_ha_2002")) as "tc_loss_ha_2002",
-              round(sum("tc_loss_ha_2003")) as "tc_loss_ha_2003",
-              round(sum("tc_loss_ha_2004")) as "tc_loss_ha_2004",
-              round(sum("tc_loss_ha_2005")) as "tc_loss_ha_2005",
-              round(sum("tc_loss_ha_2006")) as "tc_loss_ha_2006",
-              round(sum("tc_loss_ha_2007")) as "tc_loss_ha_2007",
-              round(sum("tc_loss_ha_2008")) as "tc_loss_ha_2008",
-              round(sum("tc_loss_ha_2009")) as "tc_loss_ha_2009",
-              round(sum("tc_loss_ha_2010")) as "tc_loss_ha_2010",
-              round(sum("tc_loss_ha_2011")) as "tc_loss_ha_2011",
-              round(sum("tc_loss_ha_2012")) as "tc_loss_ha_2012",
-              round(sum("tc_loss_ha_2013")) as "tc_loss_ha_2013",
-              round(sum("tc_loss_ha_2014")) as "tc_loss_ha_2014",
-              round(sum("tc_loss_ha_2015")) as "tc_loss_ha_2015",
-              round(sum("tc_loss_ha_2016")) as "tc_loss_ha_2016",
-              round(sum("tc_loss_ha_2017")) as "tc_loss_ha_2017",
-              round(sum("tc_loss_ha_2018")) as "tc_loss_ha_2018",
-              round(sum("biomass_loss_Mt_2001")) as "biomass_loss_Mt_2001",
-              round(sum("biomass_loss_Mt_2002")) as "biomass_loss_Mt_2002",
-              round(sum("biomass_loss_Mt_2003")) as "biomass_loss_Mt_2003",
-              round(sum("biomass_loss_Mt_2004")) as "biomass_loss_Mt_2004",
-              round(sum("biomass_loss_Mt_2005")) as "biomass_loss_Mt_2005",
-              round(sum("biomass_loss_Mt_2006")) as "biomass_loss_Mt_2006",
-              round(sum("biomass_loss_Mt_2007")) as "biomass_loss_Mt_2007",
-              round(sum("biomass_loss_Mt_2008")) as "biomass_loss_Mt_2008",
-              round(sum("biomass_loss_Mt_2009")) as "biomass_loss_Mt_2009",
-              round(sum("biomass_loss_Mt_2010")) as "biomass_loss_Mt_2010",
-              round(sum("biomass_loss_Mt_2011")) as "biomass_loss_Mt_2011",
-              round(sum("biomass_loss_Mt_2012")) as "biomass_loss_Mt_2012",
-              round(sum("biomass_loss_Mt_2013")) as "biomass_loss_Mt_2013",
-              round(sum("biomass_loss_Mt_2014")) as "biomass_loss_Mt_2014",
-              round(sum("biomass_loss_Mt_2015")) as "biomass_loss_Mt_2015",
-              round(sum("biomass_loss_Mt_2016")) as "biomass_loss_Mt_2016",
-              round(sum("biomass_loss_Mt_2017")) as "biomass_loss_Mt_2017",
-              round(sum("biomass_loss_Mt_2018")) as "biomass_loss_Mt_2018",
-              round(sum("carbon_emissions_Mt_2001")) as "carbon_emissions_Mt_2001",
-              round(sum("carbon_emissions_Mt_2002")) as "carbon_emissions_Mt_2002",
-              round(sum("carbon_emissions_Mt_2003")) as "carbon_emissions_Mt_2003",
-              round(sum("carbon_emissions_Mt_2004")) as "carbon_emissions_Mt_2004",
-              round(sum("carbon_emissions_Mt_2005")) as "carbon_emissions_Mt_2005",
-              round(sum("carbon_emissions_Mt_2006")) as "carbon_emissions_Mt_2006",
-              round(sum("carbon_emissions_Mt_2007")) as "carbon_emissions_Mt_2007",
-              round(sum("carbon_emissions_Mt_2008")) as "carbon_emissions_Mt_2008",
-              round(sum("carbon_emissions_Mt_2009")) as "carbon_emissions_Mt_2009",
-              round(sum("carbon_emissions_Mt_2010")) as "carbon_emissions_Mt_2010",
-              round(sum("carbon_emissions_Mt_2011")) as "carbon_emissions_Mt_2011",
-              round(sum("carbon_emissions_Mt_2012")) as "carbon_emissions_Mt_2012",
-              round(sum("carbon_emissions_Mt_2013")) as "carbon_emissions_Mt_2013",
-              round(sum("carbon_emissions_Mt_2014")) as "carbon_emissions_Mt_2014",
-              round(sum("carbon_emissions_Mt_2015")) as "carbon_emissions_Mt_2015",
-              round(sum("carbon_emissions_Mt_2016")) as "carbon_emissions_Mt_2016",
-              round(sum("carbon_emissions_Mt_2017")) as "carbon_emissions_Mt_2017",
-              round(sum("carbon_emissions_Mt_2018")) as "carbon_emissions_Mt_2018"
-            )
+          val isoSummaryDF = adm1SummaryDF.transform(IsoSummaryDF.sumArea)
 
-          isoSummaryDF.write
+          isoSummaryDF
+            .repartition($"iso")
+            .orderBy($"iso", $"threshold")
+            .write
             .options(csvOptions)
             .csv(path = runOutputUrl + "/summary/iso")
 
-
-          adm1SummaryDF.coalesce(1)
+          adm1SummaryDF
+            .coalesce(1)
+            .orderBy($"iso", $"adm1", $"threshold")
             .write
             .options(csvOptions)
             .csv(path = runOutputUrl + "/summary/global/adm1")
 
-          isoSummaryDF.coalesce(1)
+          isoSummaryDF
+            .coalesce(1)
+            .orderBy($"iso", $"threshold")
             .write
             .options(csvOptions)
             .csv(path = runOutputUrl + "/summary/global/iso")
 
           val adm2ApiDF = adm2DF
-            .select(
-              $"iso",
-              $"adm1",
-              $"adm2",
-              $"threshold",
-              $"tcs",
-              $"global_land_cover",
-              $"primary_forest",
-              $"idn_primary_forest",
-              $"erosion",
-              $"biodiversity_significance",
-              $"biodiversity_intactness",
-              $"wdpa",
-              $"aze",
-              $"plantations",
-              $"river_basin",
-              $"ecozone",
-              $"urban_watershed",
-              $"mangroves_1996",
-              $"mangroves_2016",
-              $"water_stress",
-              $"ifl",
-              $"endemic_bird_area",
-              $"tiger_cl",
-              $"landmark",
-              $"land_right",
-              $"kba",
-              $"mining",
-              $"rspo",
-              $"idn_mys_peatlands",
-              $"oil_palm",
-              $"idn_forest_moratorium",
-              $"idn_land_cover",
-              $"mex_protected_areas",
-              $"mex_pes",
-              $"mex_forest_zoning",
-              $"per_production_forest",
-              $"per_protected_area",
-              $"per_forest_concession",
-              $"bra_biomes",
-              $"wood_fiber",
-              $"resource_right",
-              $"managed_forests",
-              $"oil_gas",
-              $"totalarea" as "total_area",
-              $"extent2000" as "extent_2000",
-              $"extent2010" as "extent_2010",
-              $"total_gain",
-              $"total_biomass",
-              $"avg_biomass_per_ha",
-              $"total_co2",
-              $"total_mangrove_biomass",
-              $"avg_mangrove_biomass_per_ha",
-              $"total_mangrove_co2",
-              array(
-                struct(
-                  $"year_2001" as "year",
-                  $"area_loss_2001" as "area_loss",
-                  $"biomass_loss_2001" as "biomass_loss",
-                  $"carbon_emissions_2001" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2001" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2001" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2002" as "year",
-                  $"area_loss_2002" as "area_loss",
-                  $"biomass_loss_2002" as "biomass_loss",
-                  $"carbon_emissions_2002" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2002" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2002" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2003" as "year",
-                  $"area_loss_2003" as "area_loss",
-                  $"biomass_loss_2003" as "biomass_loss",
-                  $"carbon_emissions_2003" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2003" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2003" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2004" as "year",
-                  $"area_loss_2004" as "area_loss",
-                  $"biomass_loss_2004" as "biomass_loss",
-                  $"carbon_emissions_2004" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2004" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2004" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2005" as "year",
-                  $"area_loss_2005" as "area_loss",
-                  $"biomass_loss_2005" as "biomass_loss",
-                  $"carbon_emissions_2005" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2005" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2005" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2006" as "year",
-                  $"area_loss_2006" as "area_loss",
-                  $"biomass_loss_2006" as "biomass_loss",
-                  $"carbon_emissions_2006" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2006" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2006" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2007" as "year",
-                  $"area_loss_2007" as "area_loss",
-                  $"biomass_loss_2007" as "biomass_loss",
-                  $"carbon_emissions_2007" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2007" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2007" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2008" as "year",
-                  $"area_loss_2008" as "area_loss",
-                  $"biomass_loss_2008" as "biomass_loss",
-                  $"carbon_emissions_2008" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2008" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2008" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2009" as "year",
-                  $"area_loss_2009" as "area_loss",
-                  $"biomass_loss_2009" as "biomass_loss",
-                  $"carbon_emissions_2009" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2009" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2009" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2010" as "year",
-                  $"area_loss_2010" as "area_loss",
-                  $"biomass_loss_2010" as "biomass_loss",
-                  $"carbon_emissions_2010" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2010" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2010" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2011" as "year",
-                  $"area_loss_2011" as "area_loss",
-                  $"biomass_loss_2011" as "biomass_loss",
-                  $"carbon_emissions_2011" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2011" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2011" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2012" as "year",
-                  $"area_loss_2012" as "area_loss",
-                  $"biomass_loss_2012" as "biomass_loss",
-                  $"carbon_emissions_2012" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2012" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2012" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2013" as "year",
-                  $"area_loss_2013" as "area_loss",
-                  $"biomass_loss_2013" as "biomass_loss",
-                  $"carbon_emissions_2013" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2013" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2013" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2014" as "year",
-                  $"area_loss_2014" as "area_loss",
-                  $"biomass_loss_2014" as "biomass_loss",
-                  $"carbon_emissions_2014" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2014" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2014" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2015" as "year",
-                  $"area_loss_2015" as "area_loss",
-                  $"biomass_loss_2015" as "biomass_loss",
-                  $"carbon_emissions_2015" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2015" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2015" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2016" as "year",
-                  $"area_loss_2016" as "area_loss",
-                  $"biomass_loss_2016" as "biomass_loss",
-                  $"carbon_emissions_2016" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2016" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2016" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2017" as "year",
-                  $"area_loss_2017" as "area_loss",
-                  $"biomass_loss_2017" as "biomass_loss",
-                  $"carbon_emissions_2017" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2017" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2017" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2018" as "year",
-                  $"area_loss_2018" as "area_loss",
-                  $"biomass_loss_2018" as "biomass_loss",
-                  $"carbon_emissions_2018" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2018" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2018" as "mangrove_carbon_emissions"
-                )
-              ) as "year_data"
-            )
-          adm2ApiDF.toJSON
+            .transform(Adm2ApiDF.nestYearData)
+
+          adm2ApiDF
+            .repartition($"iso", $"adm1", $"adm2", $"threshold")
+            .orderBy($"iso", $"adm1", $"adm2", $"threshold")
+            .toJSON
             .mapPartitions(vals => Iterator("[" + vals.mkString(",") + "]"))
             .write
             .text(runOutputUrl + "/api/adm2")
 
           val adm1ApiDF = adm2DF
-            .groupBy(
-              $"iso",
-              $"adm1",
-              $"threshold",
-              $"tcs",
-              $"global_land_cover",
-              $"primary_forest",
-              $"idn_primary_forest",
-              $"erosion",
-              $"biodiversity_significance",
-              $"biodiversity_intactness",
-              $"wdpa",
-              $"aze",
-              $"plantations",
-              $"river_basin",
-              $"ecozone",
-              $"urban_watershed",
-              $"mangroves_1996",
-              $"mangroves_2016",
-              $"water_stress",
-              $"ifl",
-              $"endemic_bird_area",
-              $"tiger_cl",
-              $"landmark",
-              $"land_right",
-              $"kba",
-              $"mining",
-              $"rspo",
-              $"idn_mys_peatlands",
-              $"oil_palm",
-              $"idn_forest_moratorium",
-              $"idn_land_cover",
-              $"mex_protected_areas",
-              $"mex_pes",
-              $"mex_forest_zoning",
-              $"per_production_forest",
-              $"per_protected_area",
-              $"per_forest_concession",
-              $"bra_biomes",
-              $"wood_fiber",
-              $"resource_right",
-              $"managed_forests",
-              $"oil_gas",
-              $"year_2001",
-              $"year_2002",
-              $"year_2003",
-              $"year_2004",
-              $"year_2005",
-              $"year_2006",
-              $"year_2007",
-              $"year_2008",
-              $"year_2009",
-              $"year_2010",
-              $"year_2011",
-              $"year_2012",
-              $"year_2013",
-              $"year_2014",
-              $"year_2015",
-              $"year_2016",
-              $"year_2017",
-              $"year_2018"
-            )
-            .agg(
-              sum("totalarea") as "total_area",
-              sum("extent2000") as "extent_2000",
-              sum("extent2010") as "extent_2010",
-              sum("total_gain") as "total_gain",
-              sum("total_biomass") as "total_biomass",
-              sum($"avg_biomass_per_ha" * $"totalarea") / sum("totalarea") as "avg_biomass_per_ha",
-              sum("total_co2") as "total_co2",
-              sum("total_mangrove_biomass") as "total_mangrove_biomass",
-              sum($"avg_mangrove_biomass_per_ha" * $"totalarea") / sum("totalarea") as "avg_mangrove_biomass_per_ha",
-              sum("total_mangrove_co2") as "total_mangrove_co2",
-              sum("area_loss_2001") as "area_loss_2001",
-              sum("area_loss_2002") as "area_loss_2002",
-              sum("area_loss_2003") as "area_loss_2003",
-              sum("area_loss_2004") as "area_loss_2004",
-              sum("area_loss_2005") as "area_loss_2005",
-              sum("area_loss_2006") as "area_loss_2006",
-              sum("area_loss_2007") as "area_loss_2007",
-              sum("area_loss_2008") as "area_loss_2008",
-              sum("area_loss_2009") as "area_loss_2009",
-              sum("area_loss_2010") as "area_loss_2010",
-              sum("area_loss_2011") as "area_loss_2011",
-              sum("area_loss_2012") as "area_loss_2012",
-              sum("area_loss_2013") as "area_loss_2013",
-              sum("area_loss_2014") as "area_loss_2014",
-              sum("area_loss_2015") as "area_loss_2015",
-              sum("area_loss_2016") as "area_loss_2016",
-              sum("area_loss_2017") as "area_loss_2017",
-              sum("area_loss_2018") as "area_loss_2018",
-              sum("biomass_loss_2001") as "biomass_loss_2001",
-              sum("biomass_loss_2002") as "biomass_loss_2002",
-              sum("biomass_loss_2003") as "biomass_loss_2003",
-              sum("biomass_loss_2004") as "biomass_loss_2004",
-              sum("biomass_loss_2005") as "biomass_loss_2005",
-              sum("biomass_loss_2006") as "biomass_loss_2006",
-              sum("biomass_loss_2007") as "biomass_loss_2007",
-              sum("biomass_loss_2008") as "biomass_loss_2008",
-              sum("biomass_loss_2009") as "biomass_loss_2009",
-              sum("biomass_loss_2010") as "biomass_loss_2010",
-              sum("biomass_loss_2011") as "biomass_loss_2011",
-              sum("biomass_loss_2012") as "biomass_loss_2012",
-              sum("biomass_loss_2013") as "biomass_loss_2013",
-              sum("biomass_loss_2014") as "biomass_loss_2014",
-              sum("biomass_loss_2015") as "biomass_loss_2015",
-              sum("biomass_loss_2016") as "biomass_loss_2016",
-              sum("biomass_loss_2017") as "biomass_loss_2017",
-              sum("biomass_loss_2018") as "biomass_loss_2018",
-              sum("carbon_emissions_2001") as "carbon_emissions_2001",
-              sum("carbon_emissions_2002") as "carbon_emissions_2002",
-              sum("carbon_emissions_2003") as "carbon_emissions_2003",
-              sum("carbon_emissions_2004") as "carbon_emissions_2004",
-              sum("carbon_emissions_2005") as "carbon_emissions_2005",
-              sum("carbon_emissions_2006") as "carbon_emissions_2006",
-              sum("carbon_emissions_2007") as "carbon_emissions_2007",
-              sum("carbon_emissions_2008") as "carbon_emissions_2008",
-              sum("carbon_emissions_2009") as "carbon_emissions_2009",
-              sum("carbon_emissions_2010") as "carbon_emissions_2010",
-              sum("carbon_emissions_2011") as "carbon_emissions_2011",
-              sum("carbon_emissions_2012") as "carbon_emissions_2012",
-              sum("carbon_emissions_2013") as "carbon_emissions_2013",
-              sum("carbon_emissions_2014") as "carbon_emissions_2014",
-              sum("carbon_emissions_2015") as "carbon_emissions_2015",
-              sum("carbon_emissions_2016") as "carbon_emissions_2016",
-              sum("carbon_emissions_2017") as "carbon_emissions_2017",
-              sum("carbon_emissions_2018") as "carbon_emissions_2018",
-              sum("mangrove_biomass_loss_2001") as "mangrove_biomass_loss_2001",
-              sum("mangrove_biomass_loss_2002") as "mangrove_biomass_loss_2002",
-              sum("mangrove_biomass_loss_2003") as "mangrove_biomass_loss_2003",
-              sum("mangrove_biomass_loss_2004") as "mangrove_biomass_loss_2004",
-              sum("mangrove_biomass_loss_2005") as "mangrove_biomass_loss_2005",
-              sum("mangrove_biomass_loss_2006") as "mangrove_biomass_loss_2006",
-              sum("mangrove_biomass_loss_2007") as "mangrove_biomass_loss_2007",
-              sum("mangrove_biomass_loss_2008") as "mangrove_biomass_loss_2008",
-              sum("mangrove_biomass_loss_2009") as "mangrove_biomass_loss_2009",
-              sum("mangrove_biomass_loss_2010") as "mangrove_biomass_loss_2010",
-              sum("mangrove_biomass_loss_2011") as "mangrove_biomass_loss_2011",
-              sum("mangrove_biomass_loss_2012") as "mangrove_biomass_loss_2012",
-              sum("mangrove_biomass_loss_2013") as "mangrove_biomass_loss_2013",
-              sum("mangrove_biomass_loss_2014") as "mangrove_biomass_loss_2014",
-              sum("mangrove_biomass_loss_2015") as "mangrove_biomass_loss_2015",
-              sum("mangrove_biomass_loss_2016") as "mangrove_biomass_loss_2016",
-              sum("mangrove_biomass_loss_2017") as "mangrove_biomass_loss_2017",
-              sum("mangrove_biomass_loss_2018") as "mangrove_biomass_loss_2018",
-              sum("mangrove_carbon_emissions_2001") as "mangrove_carbon_emissions_2001",
-              sum("mangrove_carbon_emissions_2002") as "mangrove_carbon_emissions_2002",
-              sum("mangrove_carbon_emissions_2003") as "mangrove_carbon_emissions_2003",
-              sum("mangrove_carbon_emissions_2004") as "mangrove_carbon_emissions_2004",
-              sum("mangrove_carbon_emissions_2005") as "mangrove_carbon_emissions_2005",
-              sum("mangrove_carbon_emissions_2006") as "mangrove_carbon_emissions_2006",
-              sum("mangrove_carbon_emissions_2007") as "mangrove_carbon_emissions_2007",
-              sum("mangrove_carbon_emissions_2008") as "mangrove_carbon_emissions_2008",
-              sum("mangrove_carbon_emissions_2009") as "mangrove_carbon_emissions_2009",
-              sum("mangrove_carbon_emissions_2010") as "mangrove_carbon_emissions_2010",
-              sum("mangrove_carbon_emissions_2011") as "mangrove_carbon_emissions_2011",
-              sum("mangrove_carbon_emissions_2012") as "mangrove_carbon_emissions_2012",
-              sum("mangrove_carbon_emissions_2013") as "mangrove_carbon_emissions_2013",
-              sum("mangrove_carbon_emissions_2014") as "mangrove_carbon_emissions_2014",
-              sum("mangrove_carbon_emissions_2015") as "mangrove_carbon_emissions_2015",
-              sum("mangrove_carbon_emissions_2016") as "mangrove_carbon_emissions_2016",
-              sum("mangrove_carbon_emissions_2017") as "mangrove_carbon_emissions_2017",
-              sum("mangrove_carbon_emissions_2018") as "mangrove_carbon_emissions_2018"
-            )
-            .select(
-              $"iso",
-              $"adm1",
-              $"threshold",
-              $"tcs",
-              $"global_land_cover",
-              $"primary_forest",
-              $"idn_primary_forest",
-              $"erosion",
-              $"biodiversity_significance",
-              $"biodiversity_intactness",
-              $"wdpa",
-              $"aze",
-              $"plantations",
-              $"river_basin",
-              $"ecozone",
-              $"urban_watershed",
-              $"mangroves_1996",
-              $"mangroves_2016",
-              $"water_stress",
-              $"ifl",
-              $"endemic_bird_area",
-              $"tiger_cl",
-              $"landmark",
-              $"land_right",
-              $"kba",
-              $"mining",
-              $"rspo",
-              $"idn_mys_peatlands",
-              $"oil_palm",
-              $"idn_forest_moratorium",
-              $"idn_land_cover",
-              $"mex_protected_areas",
-              $"mex_pes",
-              $"mex_forest_zoning",
-              $"per_production_forest",
-              $"per_protected_area",
-              $"per_forest_concession",
-              $"bra_biomes",
-              $"wood_fiber",
-              $"resource_right",
-              $"managed_forests",
-              $"oil_gas",
-              $"total_area",
-              $"extent_2000",
-              $"extent_2010",
-              $"total_gain",
-              $"total_biomass",
-              $"avg_biomass_per_ha",
-              $"total_co2",
-              $"total_mangrove_biomass",
-              $"avg_mangrove_biomass_per_ha",
-              $"total_mangrove_co2",
-              array(
-                struct(
-                  $"year_2001" as "year",
-                  $"area_loss_2001" as "area_loss",
-                  $"biomass_loss_2001" as "biomass_loss",
-                  $"carbon_emissions_2001" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2001" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2001" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2002" as "year",
-                  $"area_loss_2002" as "area_loss",
-                  $"biomass_loss_2002" as "biomass_loss",
-                  $"carbon_emissions_2002" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2002" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2002" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2003" as "year",
-                  $"area_loss_2003" as "area_loss",
-                  $"biomass_loss_2003" as "biomass_loss",
-                  $"carbon_emissions_2003" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2003" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2003" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2004" as "year",
-                  $"area_loss_2004" as "area_loss",
-                  $"biomass_loss_2004" as "biomass_loss",
-                  $"carbon_emissions_2004" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2004" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2004" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2005" as "year",
-                  $"area_loss_2005" as "area_loss",
-                  $"biomass_loss_2005" as "biomass_loss",
-                  $"carbon_emissions_2005" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2005" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2005" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2006" as "year",
-                  $"area_loss_2006" as "area_loss",
-                  $"biomass_loss_2006" as "biomass_loss",
-                  $"carbon_emissions_2006" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2006" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2006" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2007" as "year",
-                  $"area_loss_2007" as "area_loss",
-                  $"biomass_loss_2007" as "biomass_loss",
-                  $"carbon_emissions_2007" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2007" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2007" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2008" as "year",
-                  $"area_loss_2008" as "area_loss",
-                  $"biomass_loss_2008" as "biomass_loss",
-                  $"carbon_emissions_2008" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2008" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2008" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2009" as "year",
-                  $"area_loss_2009" as "area_loss",
-                  $"biomass_loss_2009" as "biomass_loss",
-                  $"carbon_emissions_2009" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2009" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2009" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2010" as "year",
-                  $"area_loss_2010" as "area_loss",
-                  $"biomass_loss_2010" as "biomass_loss",
-                  $"carbon_emissions_2010" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2010" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2010" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2011" as "year",
-                  $"area_loss_2011" as "area_loss",
-                  $"biomass_loss_2011" as "biomass_loss",
-                  $"carbon_emissions_2011" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2011" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2011" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2012" as "year",
-                  $"area_loss_2012" as "area_loss",
-                  $"biomass_loss_2012" as "biomass_loss",
-                  $"carbon_emissions_2012" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2012" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2012" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2013" as "year",
-                  $"area_loss_2013" as "area_loss",
-                  $"biomass_loss_2013" as "biomass_loss",
-                  $"carbon_emissions_2013" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2013" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2013" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2014" as "year",
-                  $"area_loss_2014" as "area_loss",
-                  $"biomass_loss_2014" as "biomass_loss",
-                  $"carbon_emissions_2014" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2014" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2014" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2015" as "year",
-                  $"area_loss_2015" as "area_loss",
-                  $"biomass_loss_2015" as "biomass_loss",
-                  $"carbon_emissions_2015" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2015" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2015" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2016" as "year",
-                  $"area_loss_2016" as "area_loss",
-                  $"biomass_loss_2016" as "biomass_loss",
-                  $"carbon_emissions_2016" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2016" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2016" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2017" as "year",
-                  $"area_loss_2017" as "area_loss",
-                  $"biomass_loss_2017" as "biomass_loss",
-                  $"carbon_emissions_2017" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2017" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2017" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2018" as "year",
-                  $"area_loss_2018" as "area_loss",
-                  $"biomass_loss_2018" as "biomass_loss",
-                  $"carbon_emissions_2018" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2018" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2018" as "mangrove_carbon_emissions"
-                )
-              ) as "year_data"
-            )
-          adm1ApiDF.toJSON
+            .transform(Adm1ApiDF.sumArea)
+            .transform(Adm1ApiDF.nestYearData)
+
+          adm1ApiDF
+            .repartition($"iso", $"adm1", $"threshold")
+            .orderBy($"iso", $"adm1", $"threshold")
+            .toJSON
             .mapPartitions(vals => Iterator("[" + vals.mkString(",") + "]"))
             .write
             .text(runOutputUrl + "/api/adm1")
 
           val isoApiDF = adm2DF
-            .groupBy(
-              $"iso",
-              $"threshold",
-              $"tcs",
-              $"global_land_cover",
-              $"primary_forest",
-              $"idn_primary_forest",
-              $"erosion",
-              $"biodiversity_significance",
-              $"biodiversity_intactness",
-              $"wdpa",
-              $"aze",
-              $"plantations",
-              $"river_basin",
-              $"ecozone",
-              $"urban_watershed",
-              $"mangroves_1996",
-              $"mangroves_2016",
-              $"water_stress",
-              $"ifl",
-              $"endemic_bird_area",
-              $"tiger_cl",
-              $"landmark",
-              $"land_right",
-              $"kba",
-              $"mining",
-              $"rspo",
-              $"idn_mys_peatlands",
-              $"oil_palm",
-              $"idn_forest_moratorium",
-              $"idn_land_cover",
-              $"mex_protected_areas",
-              $"mex_pes",
-              $"mex_forest_zoning",
-              $"per_production_forest",
-              $"per_protected_area",
-              $"per_forest_concession",
-              $"bra_biomes",
-              $"wood_fiber",
-              $"resource_right",
-              $"managed_forests",
-              $"oil_gas",
-              $"year_2001",
-              $"year_2002",
-              $"year_2003",
-              $"year_2004",
-              $"year_2005",
-              $"year_2006",
-              $"year_2007",
-              $"year_2008",
-              $"year_2009",
-              $"year_2010",
-              $"year_2011",
-              $"year_2012",
-              $"year_2013",
-              $"year_2014",
-              $"year_2015",
-              $"year_2016",
-              $"year_2017",
-              $"year_2018"
-            )
-            .agg(
-              sum("totalarea") as "total_area",
-              sum("extent2000") as "extent_2000",
-              sum("extent2010") as "extent_2010",
-              sum("total_gain") as "total_gain",
-              sum("total_biomass") as "total_biomass",
-              sum($"avg_biomass_per_ha" * $"totalarea") / sum("totalarea") as "avg_biomass_per_ha",
-              sum("total_co2") as "total_co2",
-              sum("total_mangrove_biomass") as "total_mangrove_biomass",
-              sum($"avg_mangrove_biomass_per_ha" * $"totalarea") / sum("totalarea") as "avg_mangrove_biomass_per_ha",
-              sum("total_mangrove_co2") as "total_mangrove_co2",
-              sum("area_loss_2001") as "area_loss_2001",
-              sum("area_loss_2002") as "area_loss_2002",
-              sum("area_loss_2003") as "area_loss_2003",
-              sum("area_loss_2004") as "area_loss_2004",
-              sum("area_loss_2005") as "area_loss_2005",
-              sum("area_loss_2006") as "area_loss_2006",
-              sum("area_loss_2007") as "area_loss_2007",
-              sum("area_loss_2008") as "area_loss_2008",
-              sum("area_loss_2009") as "area_loss_2009",
-              sum("area_loss_2010") as "area_loss_2010",
-              sum("area_loss_2011") as "area_loss_2011",
-              sum("area_loss_2012") as "area_loss_2012",
-              sum("area_loss_2013") as "area_loss_2013",
-              sum("area_loss_2014") as "area_loss_2014",
-              sum("area_loss_2015") as "area_loss_2015",
-              sum("area_loss_2016") as "area_loss_2016",
-              sum("area_loss_2017") as "area_loss_2017",
-              sum("area_loss_2018") as "area_loss_2018",
-              sum("biomass_loss_2001") as "biomass_loss_2001",
-              sum("biomass_loss_2002") as "biomass_loss_2002",
-              sum("biomass_loss_2003") as "biomass_loss_2003",
-              sum("biomass_loss_2004") as "biomass_loss_2004",
-              sum("biomass_loss_2005") as "biomass_loss_2005",
-              sum("biomass_loss_2006") as "biomass_loss_2006",
-              sum("biomass_loss_2007") as "biomass_loss_2007",
-              sum("biomass_loss_2008") as "biomass_loss_2008",
-              sum("biomass_loss_2009") as "biomass_loss_2009",
-              sum("biomass_loss_2010") as "biomass_loss_2010",
-              sum("biomass_loss_2011") as "biomass_loss_2011",
-              sum("biomass_loss_2012") as "biomass_loss_2012",
-              sum("biomass_loss_2013") as "biomass_loss_2013",
-              sum("biomass_loss_2014") as "biomass_loss_2014",
-              sum("biomass_loss_2015") as "biomass_loss_2015",
-              sum("biomass_loss_2016") as "biomass_loss_2016",
-              sum("biomass_loss_2017") as "biomass_loss_2017",
-              sum("biomass_loss_2018") as "biomass_loss_2018",
-              sum("carbon_emissions_2001") as "carbon_emissions_2001",
-              sum("carbon_emissions_2002") as "carbon_emissions_2002",
-              sum("carbon_emissions_2003") as "carbon_emissions_2003",
-              sum("carbon_emissions_2004") as "carbon_emissions_2004",
-              sum("carbon_emissions_2005") as "carbon_emissions_2005",
-              sum("carbon_emissions_2006") as "carbon_emissions_2006",
-              sum("carbon_emissions_2007") as "carbon_emissions_2007",
-              sum("carbon_emissions_2008") as "carbon_emissions_2008",
-              sum("carbon_emissions_2009") as "carbon_emissions_2009",
-              sum("carbon_emissions_2010") as "carbon_emissions_2010",
-              sum("carbon_emissions_2011") as "carbon_emissions_2011",
-              sum("carbon_emissions_2012") as "carbon_emissions_2012",
-              sum("carbon_emissions_2013") as "carbon_emissions_2013",
-              sum("carbon_emissions_2014") as "carbon_emissions_2014",
-              sum("carbon_emissions_2015") as "carbon_emissions_2015",
-              sum("carbon_emissions_2016") as "carbon_emissions_2016",
-              sum("carbon_emissions_2017") as "carbon_emissions_2017",
-              sum("carbon_emissions_2018") as "carbon_emissions_2018",
-              sum("mangrove_biomass_loss_2001") as "mangrove_biomass_loss_2001",
-              sum("mangrove_biomass_loss_2002") as "mangrove_biomass_loss_2002",
-              sum("mangrove_biomass_loss_2003") as "mangrove_biomass_loss_2003",
-              sum("mangrove_biomass_loss_2004") as "mangrove_biomass_loss_2004",
-              sum("mangrove_biomass_loss_2005") as "mangrove_biomass_loss_2005",
-              sum("mangrove_biomass_loss_2006") as "mangrove_biomass_loss_2006",
-              sum("mangrove_biomass_loss_2007") as "mangrove_biomass_loss_2007",
-              sum("mangrove_biomass_loss_2008") as "mangrove_biomass_loss_2008",
-              sum("mangrove_biomass_loss_2009") as "mangrove_biomass_loss_2009",
-              sum("mangrove_biomass_loss_2010") as "mangrove_biomass_loss_2010",
-              sum("mangrove_biomass_loss_2011") as "mangrove_biomass_loss_2011",
-              sum("mangrove_biomass_loss_2012") as "mangrove_biomass_loss_2012",
-              sum("mangrove_biomass_loss_2013") as "mangrove_biomass_loss_2013",
-              sum("mangrove_biomass_loss_2014") as "mangrove_biomass_loss_2014",
-              sum("mangrove_biomass_loss_2015") as "mangrove_biomass_loss_2015",
-              sum("mangrove_biomass_loss_2016") as "mangrove_biomass_loss_2016",
-              sum("mangrove_biomass_loss_2017") as "mangrove_biomass_loss_2017",
-              sum("mangrove_biomass_loss_2018") as "mangrove_biomass_loss_2018",
-              sum("mangrove_carbon_emissions_2001") as "mangrove_carbon_emissions_2001",
-              sum("mangrove_carbon_emissions_2002") as "mangrove_carbon_emissions_2002",
-              sum("mangrove_carbon_emissions_2003") as "mangrove_carbon_emissions_2003",
-              sum("mangrove_carbon_emissions_2004") as "mangrove_carbon_emissions_2004",
-              sum("mangrove_carbon_emissions_2005") as "mangrove_carbon_emissions_2005",
-              sum("mangrove_carbon_emissions_2006") as "mangrove_carbon_emissions_2006",
-              sum("mangrove_carbon_emissions_2007") as "mangrove_carbon_emissions_2007",
-              sum("mangrove_carbon_emissions_2008") as "mangrove_carbon_emissions_2008",
-              sum("mangrove_carbon_emissions_2009") as "mangrove_carbon_emissions_2009",
-              sum("mangrove_carbon_emissions_2010") as "mangrove_carbon_emissions_2010",
-              sum("mangrove_carbon_emissions_2011") as "mangrove_carbon_emissions_2011",
-              sum("mangrove_carbon_emissions_2012") as "mangrove_carbon_emissions_2012",
-              sum("mangrove_carbon_emissions_2013") as "mangrove_carbon_emissions_2013",
-              sum("mangrove_carbon_emissions_2014") as "mangrove_carbon_emissions_2014",
-              sum("mangrove_carbon_emissions_2015") as "mangrove_carbon_emissions_2015",
-              sum("mangrove_carbon_emissions_2016") as "mangrove_carbon_emissions_2016",
-              sum("mangrove_carbon_emissions_2017") as "mangrove_carbon_emissions_2017",
-              sum("mangrove_carbon_emissions_2018") as "mangrove_carbon_emissions_2018"
-            )
-            .select(
-              $"iso",
-              $"threshold",
-              $"tcs",
-              $"global_land_cover",
-              $"primary_forest",
-              $"idn_primary_forest",
-              $"erosion",
-              $"biodiversity_significance",
-              $"biodiversity_intactness",
-              $"wdpa",
-              $"aze",
-              $"plantations",
-              $"river_basin",
-              $"ecozone",
-              $"urban_watershed",
-              $"mangroves_1996",
-              $"mangroves_2016",
-              $"water_stress",
-              $"ifl",
-              $"endemic_bird_area",
-              $"tiger_cl",
-              $"landmark",
-              $"land_right",
-              $"kba",
-              $"mining",
-              $"rspo",
-              $"idn_mys_peatlands",
-              $"oil_palm",
-              $"idn_forest_moratorium",
-              $"idn_land_cover",
-              $"mex_protected_areas",
-              $"mex_pes",
-              $"mex_forest_zoning",
-              $"per_production_forest",
-              $"per_protected_area",
-              $"per_forest_concession",
-              $"bra_biomes",
-              $"wood_fiber",
-              $"resource_right",
-              $"managed_forests",
-              $"oil_gas",
-              $"total_area",
-              $"extent_2000",
-              $"extent_2010",
-              $"total_gain",
-              $"total_biomass",
-              $"avg_biomass_per_ha",
-              $"total_co2",
-              $"total_mangrove_biomass",
-              $"avg_mangrove_biomass_per_ha",
-              $"total_mangrove_co2",
-              array(
-                struct(
-                  $"year_2001" as "year",
-                  $"area_loss_2001" as "area_loss",
-                  $"biomass_loss_2001" as "biomass_loss",
-                  $"carbon_emissions_2001" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2001" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2001" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2002" as "year",
-                  $"area_loss_2002" as "area_loss",
-                  $"biomass_loss_2002" as "biomass_loss",
-                  $"carbon_emissions_2002" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2002" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2002" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2003" as "year",
-                  $"area_loss_2003" as "area_loss",
-                  $"biomass_loss_2003" as "biomass_loss",
-                  $"carbon_emissions_2003" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2003" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2003" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2004" as "year",
-                  $"area_loss_2004" as "area_loss",
-                  $"biomass_loss_2004" as "biomass_loss",
-                  $"carbon_emissions_2004" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2004" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2004" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2005" as "year",
-                  $"area_loss_2005" as "area_loss",
-                  $"biomass_loss_2005" as "biomass_loss",
-                  $"carbon_emissions_2005" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2005" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2005" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2006" as "year",
-                  $"area_loss_2006" as "area_loss",
-                  $"biomass_loss_2006" as "biomass_loss",
-                  $"carbon_emissions_2006" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2006" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2006" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2007" as "year",
-                  $"area_loss_2007" as "area_loss",
-                  $"biomass_loss_2007" as "biomass_loss",
-                  $"carbon_emissions_2007" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2007" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2007" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2008" as "year",
-                  $"area_loss_2008" as "area_loss",
-                  $"biomass_loss_2008" as "biomass_loss",
-                  $"carbon_emissions_2008" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2008" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2008" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2009" as "year",
-                  $"area_loss_2009" as "area_loss",
-                  $"biomass_loss_2009" as "biomass_loss",
-                  $"carbon_emissions_2009" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2009" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2009" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2010" as "year",
-                  $"area_loss_2010" as "area_loss",
-                  $"biomass_loss_2010" as "biomass_loss",
-                  $"carbon_emissions_2010" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2010" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2010" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2011" as "year",
-                  $"area_loss_2011" as "area_loss",
-                  $"biomass_loss_2011" as "biomass_loss",
-                  $"carbon_emissions_2011" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2011" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2011" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2012" as "year",
-                  $"area_loss_2012" as "area_loss",
-                  $"biomass_loss_2012" as "biomass_loss",
-                  $"carbon_emissions_2012" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2012" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2012" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2013" as "year",
-                  $"area_loss_2013" as "area_loss",
-                  $"biomass_loss_2013" as "biomass_loss",
-                  $"carbon_emissions_2013" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2013" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2013" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2014" as "year",
-                  $"area_loss_2014" as "area_loss",
-                  $"biomass_loss_2014" as "biomass_loss",
-                  $"carbon_emissions_2014" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2014" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2014" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2015" as "year",
-                  $"area_loss_2015" as "area_loss",
-                  $"biomass_loss_2015" as "biomass_loss",
-                  $"carbon_emissions_2015" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2015" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2015" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2016" as "year",
-                  $"area_loss_2016" as "area_loss",
-                  $"biomass_loss_2016" as "biomass_loss",
-                  $"carbon_emissions_2016" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2016" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2016" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2017" as "year",
-                  $"area_loss_2017" as "area_loss",
-                  $"biomass_loss_2017" as "biomass_loss",
-                  $"carbon_emissions_2017" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2017" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2017" as "mangrove_carbon_emissions"
-                ),
-                struct(
-                  $"year_2018" as "year",
-                  $"area_loss_2018" as "area_loss",
-                  $"biomass_loss_2018" as "biomass_loss",
-                  $"carbon_emissions_2018" as "carbon_emissions",
-                  $"mangrove_biomass_loss_2018" as "mangrove_biomass_loss",
-                  $"mangrove_carbon_emissions_2018" as "mangrove_carbon_emissions"
-                )
-              ) as "year_data"
-            )
-          isoApiDF.toJSON
+            .transform(IsoApiDF.sumArea)
+            .transform(IsoApiDF.nestYearData)
+
+          isoApiDF
+            .repartition($"iso", $"threshold")
+            .orderBy($"iso", $"threshold")
+            .toJSON
             .mapPartitions(vals => Iterator("[" + vals.mkString(",") + "]"))
             .write
             .text(runOutputUrl + "/api/iso")
